Split control.rs, control_p2.rs, control_p3.rs at clean impl-block boundaries to clear max-lines
