+++ crates/almide-mir/src/lower/control_p2.rs
@@ -635,1993 +635,4 @@ impl LowerCtx {
         }
         self.heap_elem_lists.insert(subj);
     }
-
-    /// The SUBJECT phase of [`Self::try_lower_variant_value_match`]: materialize/
-    /// borrow + track the match subject (effect-result, self-host call, user call,
-    /// member, tracked var …) and classify its Option/Result repr. Returns `None`
-    /// AFTER rolling back to the given marks (the caller's rollback discipline).
-    /// Verbatim text move (#781).
-    fn variant_match_subject(
-        &mut self,
-        subject: &IrExpr,
-        ops_mark: usize,
-        lhh_mark: usize,
-    ) -> Option<(ValueId, bool, bool, bool)> {
-        // Roll back PERSISTENT side effects too: a probed subject containing a lambda
-        // LIFTS it (self.lifted), and an abandoned probe would leave a dead lifted fn
-        // whose CallFn ops double-count the caps gate's mir tally (mir > ir breach).
-        let lifted_mark = self.lifted.len();
-        let rollback = |s: &mut Self| {
-            s.ops.truncate(ops_mark);
-            s.lifted.truncate(lifted_mark);
-            s.live_heap_handles.truncate(lhh_mark);
-            None
-        };
-        // EFFECT-RESULT SUBJECT (#76): a ctor-`match` over a CAN-ERR EFFECT call returning a Result
-        // — an `@intrinsic`/impure stdlib `Module` effect (`process.kill`/`process.spawn`) or a bare
-        // effect `RuntimeCall`. `lower_call_args` REFUSES such a heap-result effect call in argument
-        // position (it would defer to an empty Opaque), so the match over it walled. Materialize it
-        // HERE as a real OWNED Result handle so the tag-read below executes. HOLE-1: gated on
-        // `effect_unwrap_admitted` (Result whose Ok payload has a real recursive drop — scalar /
-        // String / Value / List[Value] / tuple-Ok); a RECORD-Ok effect result is REFUSED here and
-        // falls through to the ordinary path (which walls), NEVER routed through a leaky flat cert.
-        let used_effect_subj = self.is_effect_result_subject(subject);
-        // Materialize/borrow + track the subject exactly as the statement Match entry does:
-        // an owned ctor temp (`Some(5)`) is dropped at scope end; a tracked Var (`let o =
-        // Some(5)`) is borrowed; a self-host Option/Result-returning call is tracked here.
-        let subj = if used_effect_subj {
-            match self.try_materialize_effect_result_subject(subject) {
-                Some(v) => v,
-                None => return rollback(self),
-            }
-        } else {
-            match self
-            .lower_call_args(std::slice::from_ref(subject))
-            .ok()
-            .and_then(|a| a.into_iter().next())
-        {
-            Some(CallArg::Handle(v)) => v,
-            _ => return rollback(self),
-        }
-        };
-        // A USER-FN `Named` call returning Option/Result is tracked the SAME as a self-host call: every
-        // Option/Result value uses the one DynListStr len-as-tag repr (brick #51), so `match find_colon(t)
-        // { none => …, some(cp) => … }` over `fn find_colon(..) -> Option[Int]` reads the tag @4 + binds the
-        // scalar payload @12 identically. `subj` is the OWNED call result (live, dropped-before for a scalar
-        // payload), and the tracking is per-subject. A HEAP-Ok user Result still self-gates (heap_or_scalar_
-        // bind requires a str-result), so only scalar payloads lower — never a silently-wrong heap move-out.
-        let is_named_call =
-            matches!(&subject.kind, IrExprKind::Call { target: CallTarget::Named { .. }, .. });
-        if is_self_host_option_call(subject)
-            || (is_named_call
-                && crate::lower::is_variant_ty(&subject.ty)
-                && !crate::lower::is_result_ty(&subject.ty))
-        {
-            self.materialized_options.insert(subj);
-            // An `Option[heap]` (`list.first(path): Option[String]` — toml set_nested's
-            // `match list.first(path)`) OWNS its payload: track it as a nested-ownership list so the
-            // Some-payload bind reads the borrowed element handle AND the scope-end drop is the
-            // recursive DropListStr (mirrors control.rs:100 for the statement-match path).
-            if crate::lower::is_heap_elem_list_ty(&subject.ty) {
-                self.heap_elem_lists.insert(subj);
-            }
-        }
-        if (is_self_host_result_call(subject) && !Self::is_heap_ok_result(&subject.ty))
-            || (is_named_call
-                && crate::lower::is_result_ty(&subject.ty)
-                && !Self::is_heap_ok_result(&subject.ty))
-            // An EFFECT-result subject (process.kill / RuntimeCall) with a SCALAR-Ok / heap-Err
-            // Result is tracked the SAME as a scalar self-host/Named result: len-as-tag @4, Err arm
-            // binds slot-0 String, subject drops via DropListStr (the case-A heap_elem_lists below).
-            // The self-host list is guarded by TYPE too: `result.collect`/`result.map` are listed
-            // there but return a HEAP-Ok Result for heap payloads — cap-as-tag, NOT len-as-tag —
-            // which the str-result branch below now tracks (reading @4 misdispatched EVERY heap-Ok
-            // collect to the Err arm — the parse_all List[Int]-as-List[String] garbage join).
-            || (used_effect_subj && !Self::is_heap_ok_result(&subject.ty))
-        {
-            self.materialized_results.insert(subj);
-            // Camp-4 sub-case 1 — a SCALAR-Ok / HEAP-Err `Result[Int, String]` (char_to_val; the
-            // unwrap-`!`-desugar's `err($x) => err($x)` re-wrap). The len-as-tag read stays @4
-            // (materialized_results, NOT _str), but ALSO track heap_elem_lists so (a) the Err arm's
-            // String payload bind is ADMITTED (heap_or_scalar_bind) and (b) the subject drops via
-            // `DropListStr` — EXACTLY right for this layout: Ok=len0 frees nothing (the int is scalar),
-            // Err=len1 frees slot-0's String. Gated to scalar-Ok + heap-Err so a heap-Ok Result (a
-            // different layout) is untouched. The Err arm move-out auto-`Dup`s in lower_heap_result_arm,
-            // so drop-subject-after frees slot-0 once (no double-free — gate-checked).
-            if let Ty::Applied(almide_lang::types::constructor::TypeConstructorId::Result, a) = &subject.ty {
-                if a.len() == 2 && !is_heap_ty(&a[0]) && is_heap_ty(&a[1]) {
-                    self.heap_elem_lists.insert(subj);
-                }
-            }
-        }
-        // A self-host HEAP-Ok Result (`value.as_string`/`value.as_array`/`result.zip` — cap-as-tag
-        // DynListStr) is tracked as a str-result (the match reads tag @16 + binds the @12 payload
-        // handle). The DROP differs by Ok-arm type: a `List[Value]` Ok (`value.as_array`) frees
-        // RECURSIVELY (`value_result_lists` → `DropResultListValue`), else a String Ok frees flat
-        // (`heap_elem_lists` → `DropListStr`). Type-driven so it is sound at every tracking site.
-        // Camp-4 sub-case 2 — a USER heap-Ok `Result[heap, String]` (decode_chunks's
-        // `Result[List[Int], String]`). Its CONSTRUCTION goes through `materialize_result_str`
-        // (cap-as-tag @16, slot-0 @12 = the heap payload), so the MATCH must read cap-tag @16 too —
-        // route it through the SAME str-result tracking (materialized_results_str + the by-type drop
-        // below: List[Value]→value_result_lists, Value→value_result_results, else flat→heap_elem_lists
-        // = DropListStr, exact for a List[Int]/String Ok). WITHOUT this the match read the len-as-tag
-        // @4 over a cap-tag block — a SILENT MISCOMPILE (Ok payload + tag both misread).
-        // *** HOLE-1 (the gate-INVISIBLE leak, now CLOSED by a real recursive drop) *** A record-Ok
-        // subject (`Result[<recursive-drop record>, String]` — resolve_run_caps' `Result[Manifest,
-        // String]`, load_manifest's nested record-Ok) is now ADMITTED: it carries the SAME cap-as-tag
-        // DynListStr repr as every other str-result (slot-0 @12 = the Ok record / Err String, tag @16),
-        // so the match reads tag @16 + binds the @12 payload identically. The ONE thing that made it a
-        // leak — the flat `else => heap_elem_lists` (DropListStr) freeing only the @12 record HANDLE and
-        // LEAKING the record's nested heap fields — is replaced below by routing the subject's scope-end
-        // drop through `variant_drop_handles="resrec:<R>"` → `Op::DropWrapperRec { is_result: true }`:
-        // at the wrapper's last ref it recurses into the @12 record via the generated `$__drop_<R>` (Ok
-        // tag) / `rc_dec`s the @12 Err String, then frees the wrapper — freeing every nested heap field
-        // exactly once (the 6625a5d3 / f75eecae machinery, the SAME `resrec:` the record-Ok CONSTRUCTION
-        // side already uses via `try_lower_result_record_ctor`). Gated on `result_ok_record_drop_fn`
-        // (the record HAS a generated `$__drop_<R>`); a record without one keeps the sound flat path.
-        if is_self_host_result_str_call(subject)
-            || (is_named_call && Self::is_heap_ok_result(&subject.ty))
-            // An EFFECT-result subject with a HEAP-Ok Result (String/Value/List[Value]/tuple-Ok —
-            // `effect_unwrap_admitted` already excluded RECORD-Ok; the by-type dispatch below is exact).
-            || (used_effect_subj && Self::is_heap_ok_result(&subject.ty))
-            // Any OTHER self-host Module call with a HEAP-Ok Result (`result.collect` /
-            // `result.map` over a heap payload — listed len-as-tag but cap-as-tag for these
-            // instantiations): TYPE decides the repr, not the list. Every heap-Ok Result is
-            // BUILT cap-as-tag (the ok()/err() ctors' materialize_result_str layout), so the
-            // read side must agree universally.
-            || (is_self_host_result_call(subject) && Self::is_heap_ok_result(&subject.ty))
-        {
-            self.materialized_results_str.insert(subj);
-            self.track_heap_ok_result_subject_drop(subj, &subject.ty);
-        }
-        // Dispatch on the tracking set. An Option reads len-as-tag (Some=len≠0); a scalar
-        // Result reads len-as-tag INVERSE (Err=len≠0, Ok=len0). The if-skeleton is uniform
-        // (then = tag≠0, else = tag==0): Option → then=Some/else=None; Result → then=Err/else=Ok.
-        // A BORROWED Option[heap] / Result[heap] FIELD subject (`match u.email { some(e) => …,
-        // none => … }` where `email: Option[String]` is a field of the param `u`). `lower_call_args`
-        // borrowed the field's variant handle — a nested-ownership block `u` still OWNS (freed by
-        // `u`'s drop; NOT in `live_heap_handles`, so no owned-subject-drop conflict). Track it so the
-        // tag-read + heap-payload BORROW bind below execute (the some-arm `e` is a second borrow of
-        // the Option's @12 slot; the result String is moved out fresh). Same len-as-tag layout as a
-        // self-host Option/Result value — only the SOURCE (a field borrow, not a call) differs.
-        if matches!(&subject.kind, IrExprKind::Member { .. } | IrExprKind::TupleIndex { .. }) {
-            use almide_lang::types::constructor::TypeConstructorId;
-            match &subject.ty {
-                Ty::Applied(TypeConstructorId::Option, a) if a.len() == 1 => {
-                    self.materialized_options.insert(subj);
-                    if is_heap_ty(&a[0]) {
-                        self.heap_elem_lists.insert(subj);
-                    }
-                }
-                Ty::Applied(TypeConstructorId::Result, a)
-                    if a.len() == 2 && !Self::is_heap_ok_result(&subject.ty) =>
-                {
-                    self.materialized_results.insert(subj);
-                    if is_heap_ty(&a[1]) {
-                        self.heap_elem_lists.insert(subj);
-                    }
-                }
-                _ => {}
-            }
-        }
-        let is_option = self.materialized_options.contains(&subj);
-        // A scalar Result reads len-as-tag (@4); a HEAP-Ok `Result[String,String]` (value.as_string,
-        // the cap-as-tag DynListStr) reads the tag at the slot-0 HIGH 32 bits (@16). Both arrange
-        // Err=then(tag≠0)/Ok=else(tag0); only the tag OFFSET differs. A str-result match here is
-        // ADMITTED for WILDCARD/scalar binds (`match value.as_string(v) { ok(_) => …, err(_) => … }`
-        // — is_scalar_type); a heap-payload bind over a str-result (`ok(s: String)`) is the Camp-4
-        // borrowed-slot case → gated out below (heap_or_scalar_bind already requires heap_elem_lists,
-        // and the heap-RESULT branch defers it).
-        let is_result_str = self.materialized_results_str.contains(&subj);
-        let is_result = self.materialized_results.contains(&subj) || is_result_str;
-        if !is_option && !is_result {
-            return rollback(self);
-        }
-        Some((subj, is_option, is_result_str, is_result))
-    }
-
-    /// If `ty` is a CUSTOM variant (a user ADT) with a registered [`VariantLayout`], return its
-    /// type name. Handles the three surface forms a variant type takes
-    /// (`Named` / inline `Variant` / `Applied(UserDefined)`). `None` for Option/Result (those
-    /// use the dedicated len-as-tag path) and every non-variant type.
-    /// Is `subject` a Result-returning call `lower_call_args` REFUSES in subject position (it would
-    /// defer to an empty Opaque, walling the ctor-match over it), but whose Ok payload the ctor-match
-    /// CAN bind/drop soundly? The admitted Ok set is `effect_unwrap_admitted` (scalar / String /
-    /// Value / List[Value] / tuple-Ok) PLUS a RECORD-Ok with a generated `$__drop_<R>`
-    /// (`result_ok_record_drop_fn` — the HOLE-1 record-result now has a real recursive drop). Three
-    /// call shapes materialize via [`Self::try_materialize_effect_result_subject`]:
-    ///   1. an IMPURE stdlib/cross-module `Module` effect (`process.kill`, `manifest.load_manifest`) —
-    ///      the original #76 case (CallFn the dotted effect name);
-    ///   2. a bare effect `RuntimeCall`;
-    ///   3. a PURE heap-Result `Module` call (`json.parse` / `toml.parse`) — the `let x = parse(c)!`
-    ///      monadic-desugar subject. It is NOT a tracked self-host option/result call (those keep
-    ///      their existing borrow/track path through `lower_call_args`), so it walled; route it
-    ///      through `lower_pure_module_value_call` (the SAME emitted CallFn name as every other call
-    ///      site — `list_heap_call_name`, never a raw dotted name) so the match reads a real block.
-    pub(crate) fn is_effect_result_subject(&self, subject: &IrExpr) -> bool {
-        if !effect_unwrap_admitted(&subject.ty, &self.variant_layouts)
-            && self.result_ok_record_drop_fn(&subject.ty).is_none()
-        {
-            return false;
-        }
-        match &subject.kind {
-            IrExprKind::Call { target: CallTarget::Module { module, func, .. }, .. } => {
-                !crate::purity::is_pure(module.as_str(), func.as_str())
-                    // A PURE heap-Result module call NOT already handled as a self-host
-                    // option/result subject (those route through `lower_call_args` + the existing
-                    // borrow/track sets, unchanged).
-                    || (!is_self_host_option_call(subject)
-                        && !is_self_host_result_call(subject)
-                        && !is_self_host_result_str_call(subject))
-            }
-            IrExprKind::RuntimeCall { .. } => true,
-            _ => false,
-        }
-    }
-
-    /// Materialize a Result-call subject ([`Self::is_effect_result_subject`]) into a real OWNED
-    /// Result handle pushed to `live_heap_handles` so the SUBJECT-DROP-BEFORE-ARMS / drop-after
-    /// machinery frees it exactly once. A PURE `Module` call routes through
-    /// `lower_pure_module_value_call` (its CallFn name = `list_heap_call_name`, byte-identical to
-    /// every other site, and it records the call's caps properly — a pure `json.parse` is
-    /// Stdout-free). An IMPURE `Module` effect / a `RuntimeCall` emit a direct `Op::CallFn` with the
-    /// dotted/symbol name (the #76 path). Self-rolls-back on any partial failure. The OWNERSHIP cert
-    /// stays exact (the heap result backs one `i`).
-    pub(crate) fn try_materialize_effect_result_subject(&mut self, subject: &IrExpr) -> Option<ValueId> {
-        let ops_mark = self.ops.len();
-        let lifted_mark = self.lifted.len();
-        let lhh_mark = self.live_heap_handles.len();
-        let rollback = |s: &mut Self| {
-            s.ops.truncate(ops_mark);
-            s.lifted.truncate(lifted_mark);
-            s.live_heap_handles.truncate(lhh_mark);
-            None
-        };
-        // PURE module call (json.parse / toml.parse): route through the standard pure-module value
-        // call so the emitted CallFn name + arg lowering match every other site.
-        if let IrExprKind::Call { target: CallTarget::Module { module, func, .. }, args, .. } =
-            &subject.kind
-        {
-            if crate::purity::is_pure(module.as_str(), func.as_str()) {
-                let Ok(dst) = self.lower_pure_module_value_call(
-                    module.as_str(),
-                    func.as_str(),
-                    args,
-                    &subject.ty,
-                ) else {
-                    return rollback(self);
-                };
-                if !self.live_heap_handles.contains(&dst) {
-                    self.live_heap_handles.push(dst);
-                }
-                return Some(dst);
-            }
-        }
-        // IMPURE effect Module / RuntimeCall: CallFn the dotted/symbol name directly.
-        let (name, args): (String, &[IrExpr]) = match &subject.kind {
-            IrExprKind::Call { target: CallTarget::Module { module, func, .. }, args, .. } => {
-                (format!("{}.{}", module.as_str(), func.as_str()), args)
-            }
-            IrExprKind::RuntimeCall { symbol, args } => (symbol.as_str().to_string(), args),
-            _ => return None,
-        };
-        let Ok(lowered) = self.lower_call_args(args) else {
-            return rollback(self);
-        };
-        let Ok(repr) = repr_of(&subject.ty) else {
-            return rollback(self);
-        };
-        let dst = self.fresh_value();
-        self.ops.push(Op::CallFn { dst: Some(dst), name, args: lowered, result: Some(repr) });
-        self.live_heap_handles.push(dst);
-        Some(dst)
-    }
-
-    /// Is the Map KEY type (of the first-arg/result Map) a NULLARY-ONLY variant
-    /// (every case fieldless — `Direction`)? Gates the `_vtag` tag-normalized map
-    /// family in `list_heap_call_name` (a free fn without layout access).
-    pub(crate) fn map_key_is_nullary_variant(&self, arg_tys: &[Ty], result_ty: &Ty) -> bool {
-        use almide_lang::types::constructor::TypeConstructorId;
-        let probe = |t: &Ty| {
-            matches!(t, Ty::Applied(TypeConstructorId::Map, a)
-                if a.len() == 2
-                    && self
-                        .custom_variant_type_name(&a[0])
-                        .and_then(|n| self.variant_layouts.by_type.get(&n).cloned())
-                        .is_some_and(|l| {
-                            !l.cases.is_empty() && l.cases.iter().all(|c| c.fields.is_empty())
-                        }))
-        };
-        arg_tys.first().is_some_and(probe) || probe(result_ty)
-    }
-
-    /// Is the Map KEY type an ALL-Int/Bool-field record (`Color { r, g, b }`)?
-    /// Gates the `_srec` string-normalized map family (a Float field's
-    /// bits-to-string would split -0.0/NaN from native's f64 eq — excluded).
-    pub(crate) fn map_key_is_scalar_record(&self, arg_tys: &[Ty], result_ty: &Ty) -> bool {
-        use almide_lang::types::constructor::TypeConstructorId;
-        let probe = |t: &Ty| {
-            matches!(t, Ty::Applied(TypeConstructorId::Map, a)
-                if a.len() == 2
-                    && self.custom_variant_type_name(&a[0]).is_none()
-                    && self
-                        .aggregate_field_tys(&a[0])
-                        .is_some_and(|(_, tys)| {
-                            !tys.is_empty()
-                                && tys.iter().all(|t| matches!(t, Ty::Int | Ty::Bool))
-                        }))
-        };
-        arg_tys.first().is_some_and(probe) || probe(result_ty)
-    }
-
-    /// The C-015 STRING-FIELD-record key/element intercept: `Map[P, <scalar|String>]`
-    /// / `Set[P]` / `list.unique(List[P])` where P is a declared record of String +
-    /// Int/Bool fields (≥1 String — the all-scalar case is the `_srec` family's) —
-    /// routed to the GENERATED `__krec_*` twins (drop_sources.rs): the key normalizes
-    /// INJECTIVELY into a String (length-prefixed content), and the backing container
-    /// is the proven `_str`/`_skv` family. Lookup/build fns only — an
-    /// iteration-returning fn would surface the normalized strings and keeps its wall.
-    pub(crate) fn krec_call_name(
-        &self,
-        module: &str,
-        func: &str,
-        arg_tys: &[Ty],
-        result_ty: &Ty,
-    ) -> Option<String> {
-        use almide_lang::types::constructor::TypeConstructorId;
-        let strrec = |t: &Ty| -> Option<String> {
-            let Ty::Named(n, _) = t else { return None };
-            if self.custom_variant_type_name(t).is_some() {
-                return None;
-            }
-            let (_, tys) = self.aggregate_field_tys(t)?;
-            (!tys.is_empty()
-                && tys.iter().all(|f| matches!(f, Ty::Int | Ty::Bool | Ty::String))
-                && tys.iter().any(|f| matches!(f, Ty::String)))
-            .then(|| n.as_str().to_string())
-        };
-        match module {
-            "map" => {
-                let probe = |t: &Ty| -> Option<(String, Ty)> {
-                    let Ty::Applied(TypeConstructorId::Map, a) = t else { return None };
-                    if a.len() != 2 {
-                        return None;
-                    }
-                    Some((strrec(&a[0])?, a[1].clone()))
-                };
-                let (rname, vty) =
-                    arg_tys.first().and_then(&probe).or_else(|| probe(result_ty))?;
-                let vclass = match vty {
-                    Ty::Int | Ty::Bool => "iv",
-                    Ty::String => "sv",
-                    _ => return None,
-                };
-                match func {
-                    "from_list" | "set" | "get" | "contains" => {
-                        Some(format!("__krec_map_{func}_{rname}_{vclass}"))
-                    }
-                    // len reads the BACKING flavor's header directly.
-                    "len" => Some(if vclass == "iv" {
-                        "map.len_skv".to_string()
-                    } else {
-                        "map.len_str".to_string()
-                    }),
-                    _ => None,
-                }
-            }
-            "set" => {
-                let probe = |t: &Ty| -> Option<String> {
-                    let Ty::Applied(TypeConstructorId::Set, a) = t else { return None };
-                    if a.len() != 1 {
-                        return None;
-                    }
-                    strrec(&a[0])
-                };
-                let rname = arg_tys.first().and_then(&probe).or_else(|| probe(result_ty))?;
-                match func {
-                    "from_list" | "contains" | "insert" => {
-                        Some(format!("__krec_set_{func}_{rname}"))
-                    }
-                    _ => None,
-                }
-            }
-            "list" if func == "unique" => {
-                let Some(Ty::Applied(TypeConstructorId::List, a)) = arg_tys.first() else {
-                    return None;
-                };
-                if a.len() != 1 {
-                    return None;
-                }
-                // An UNANNOTATED record literal keeps its STRUCTURAL type (the r5
-                // lesson) — its block lays fields in SOURCE order, so the norm is
-                // generated against the structural shape, keyed by the anon hash.
-                if let Ty::Record { fields } = &a[0] {
-                    if !fields.is_empty()
-                        && fields
-                            .iter()
-                            .all(|(_, t)| matches!(t, Ty::Int | Ty::Bool | Ty::String))
-                        && fields.iter().any(|(_, t)| matches!(t, Ty::String))
-                    {
-                        return Some(format!(
-                            "__krec_list_unique_{}",
-                            crate::lower::anon_record_drop_name(fields)
-                        ));
-                    }
-                    return None;
-                }
-                let rname = strrec(&a[0])?;
-                Some(format!("__krec_list_unique_{rname}"))
-            }
-            _ => None,
-        }
-    }
-
-    pub(crate) fn custom_variant_type_name(&self, ty: &Ty) -> Option<String> {
-        use almide_lang::types::constructor::TypeConstructorId;
-        let name = match ty {
-            Ty::Named(n, _) => n.as_str().to_string(),
-            Ty::Variant { name, .. } => name.as_str().to_string(),
-            Ty::Applied(TypeConstructorId::UserDefined(n), _) => n.clone(),
-            _ => return None,
-        };
-        self.variant_layouts.by_type.contains_key(&name).then_some(name)
-    }
-
-    /// EXECUTE a `match v { Ctor(binds…) => arm, … }` over a CUSTOM variant (ADT brick 3) —
-    /// read the tag from slot 0 and dispatch to the matching arm; only the taken arm runs. The
-    /// N-constructor generalization of [`Self::try_lower_variant_value_match`] (the 2-variant
-    /// Option/Result case), over the v1 tag@slot0 + i64-slot value model (NOT the len-as-tag
-    /// Option/Result repr). Returns the scalar result `dst`, or `None` (rolled back) outside
-    /// the subset — the caller then walls (a Const-0 would silently pick a wrong arm — the ②
-    /// cardinal rule).
-    ///
-    /// SUBSET: SCALAR result (ADT brick 3) OR a HEAP result over a BORROWED subject (ADT brick 4,
-    /// e.g. recursive `to_string` — each arm reads the borrowed subject's scalar slots and moves
-    /// out a fresh heap value). SCALAR ctor-field binds only (a heap/nested ctor field = ADT
-    /// brick 5). No guards. An OWNED-temp subject with a heap result would need
-    /// subject-drop-before-arms (the cert rejects the owned-borrow / arm-move-out overlap) — it
-    /// WALLS here rather than emit cert-failing MIR (ADT brick 4b).
-    ///
-    /// SOUNDNESS: the subject is materialized/borrowed by `lower_call_args` (an owned ctor temp
-    /// drops at scope end via `live_heap_handles`; a tracked Var/param borrows). The tag/field
-    /// reads are scalar prims, the `IfThen`/`Else`/`EndIf` markers no-op in `verify_ownership`,
-    /// and each arm is a per-arm-balanced frame (`lower_scalar_arm` / `lower_heap_result_arm`)
-    /// with NO heap ownership event beyond the arm's own move-out — exactly the per-arm
-    /// linearization the cert proves, wrapped so one arm runs. The LAST arm is the unconditional
-    /// `else` (the frontend guarantees the match is exhaustive).
-    /// TAIL-VALUE match over an `Option[<heap>]` subject with HEAP-result arms — the
-    /// Option twin of [`Self::try_lower_result_match_value`] (the shape its scalar-payload
-    /// self-gate called "the true Camp-4 frontier"):
-    ///   `match acc { none => none, some(stack) => <heap arm> }`  (is_balanced's fold step)
-    /// Same merge discipline as the Result opener: subject via `lower_call_args` (a borrowed
-    /// param stays caller-owned; an owned temp joins the scope epilogue), tag = len@4 with
-    /// OPTION polarity (THEN tag≠0 = Some, ELSE = None), the Some payload bound as a
-    /// slot-0 HANDLE BORROW (`param_values`, not a second owner — `lower_heap_result_arm`
-    /// Dups any borrowed payload it re-wraps), arms via `lower_heap_result_arm` with the
-    /// release-parity sweep.
-    pub(crate) fn try_lower_option_match_value(
-        &mut self,
-        subject: &IrExpr,
-        arms: &[IrMatchArm],
-        result_ty: &Ty,
-    ) -> Option<ValueId> {
-        use crate::PrimKind;
-        use almide_lang::types::constructor::TypeConstructorId;
-        if !is_heap_ty(result_ty) || arms.len() != 2 || arms.iter().any(|a| a.guard.is_some()) {
-            return None;
-        }
-        // HEAP payloads only — a scalar payload already executes via
-        // `try_lower_variant_value_match` (tried before this in every caller).
-        match &subject.ty {
-            Ty::Applied(TypeConstructorId::Option, a) if a.len() == 1 && is_heap_ty(&a[0]) => {}
-            _ => return None,
-        }
-        let mut some_arm: Option<(&IrExpr, Option<VarId>)> = None;
-        let mut none_arm: Option<&IrExpr> = None;
-        for arm in arms {
-            match &arm.pattern {
-                IrPattern::Some { inner } => {
-                    let bind = match inner.as_ref() {
-                        IrPattern::Bind { var, .. } => Some(*var),
-                        IrPattern::Wildcard => None,
-                        _ => return None,
-                    };
-                    if some_arm.is_some() {
-                        return None;
-                    }
-                    some_arm = Some((&arm.body, bind));
-                }
-                IrPattern::None => {
-                    if none_arm.is_some() {
-                        return None;
-                    }
-                    none_arm = Some(&arm.body);
-                }
-                IrPattern::Wildcard => {
-                    if none_arm.is_some() {
-                        return None;
-                    }
-                    none_arm = Some(&arm.body);
-                }
-                _ => return None,
-            }
-        }
-        let ((some_body, some_bind), none_body) = match (some_arm, none_arm) {
-            (Some(s), Some(n)) => (s, n),
-            _ => return None,
-        };
-        let ops_mark = self.ops.len();
-        let lifted_mark = self.lifted.len();
-        let lhh_mark = self.live_heap_handles.len();
-        let subj = match self
-            .lower_call_args(std::slice::from_ref(subject))
-            .ok()
-            .and_then(|a| a.into_iter().next())
-        {
-            Some(CallArg::Handle(v)) => v,
-            _ => {
-                self.ops.truncate(ops_mark);
-                self.lifted.truncate(lifted_mark);
-                self.live_heap_handles.truncate(lhh_mark);
-                return None;
-            }
-        };
-        if self.deferred_opaque_binds.contains(&subj) {
-            self.ops.truncate(ops_mark);
-            self.lifted.truncate(lifted_mark);
-            self.live_heap_handles.truncate(lhh_mark);
-            return None;
-        }
-        let h = self.fresh_value();
-        self.ops.push(Op::Prim { kind: PrimKind::Handle, dst: Some(h), args: vec![subj] });
-        let tag = self.load_at_offset(h, 4, PrimKind::Load { width: 4 });
-        let dst = self.fresh_value();
-        self.ops.push(Op::IfThen { cond: tag, dst: Some(dst) });
-        // THEN (tag != 0 = Some): payload = the slot-0 HANDLE, a BORROW of the subject's slot.
-        if let Some(var) = some_bind {
-            let payload = self.load_at_offset(h, 12, PrimKind::LoadHandle);
-            self.param_values.insert(payload);
-            self.value_of.insert(var, payload);
-        }
-        let outer: Vec<ValueId> = self.live_heap_handles.clone();
-        let some_obj = match self.lower_heap_result_arm(some_body, result_ty) {
-            Some(v) => v,
-            _ => {
-                self.ops.truncate(ops_mark);
-                self.lifted.truncate(lifted_mark);
-                self.live_heap_handles.truncate(lhh_mark);
-                return None;
-            }
-        };
-        let consumed_by_some: Vec<ValueId> =
-            outer.iter().copied().filter(|x| !self.live_heap_handles.contains(x)).collect();
-        let else_marker_at = self.ops.len();
-        self.ops.push(Op::Else { val: Some(some_obj) });
-        // ELSE (tag == 0 = None).
-        let live_after_some: Vec<ValueId> = self.live_heap_handles.clone();
-        let none_obj = match self.lower_heap_result_arm(none_body, result_ty) {
-            Some(v) => v,
-            _ => {
-                self.ops.truncate(ops_mark);
-                self.lifted.truncate(lifted_mark);
-                self.live_heap_handles.truncate(lhh_mark);
-                return None;
-            }
-        };
-        let consumed_by_none: Vec<ValueId> = live_after_some
-            .iter()
-            .copied()
-            .filter(|x| !self.live_heap_handles.contains(x))
-            .collect();
-        // Release parity across the arms (the lower_heap_result_if_inner discipline).
-        for x in &consumed_by_some {
-            if !consumed_by_none.contains(x) {
-                let op = self.drop_op_for(*x);
-                self.ops.push(op);
-            }
-        }
-        for x in &consumed_by_none {
-            if !consumed_by_some.contains(x) {
-                let op = self.drop_op_for(*x);
-                self.ops.insert(else_marker_at, op);
-            }
-        }
-        self.ops.push(Op::EndIf { val: Some(none_obj) });
-        Some(dst)
-    }
-
-    /// TAIL-VALUE Result match over a len-as-tag subject with HEAP-result arms — the
-    /// Camp-4 opener for the `compute` class:
-    ///   `match safe_div(a, b) { ok(v) => ok(int.to_string(v)), err(e) => <heap arm> }`
-    /// The SUBJECT is materialized as an OWNED tracked temp (a call result / Dup'd var —
-    /// dropped at scope end, AFTER the arms: each arm binds its payload as a BORROW (a
-    /// scalar copy for Ok @12; the slot-0 HANDLE for a heap Err — `param_values`, not a
-    /// second owner) and constructs its own FRESH result via `lower_heap_result_arm`
-    /// (which Dups any borrowed payload it re-wraps), so nothing outlives the subject.
-    /// The IfThen/Else/EndIf merge carries the arm value out — the released-merge cert
-    /// shape `lower_heap_result_if_inner` already proves, incl. the release-parity sweep.
-    pub(crate) fn try_lower_result_match_value(
-        &mut self,
-        subject: &IrExpr,
-        arms: &[IrMatchArm],
-        result_ty: &Ty,
-    ) -> Option<ValueId> {
-        use crate::PrimKind;
-        use almide_lang::types::constructor::TypeConstructorId;
-        if !is_heap_ty(result_ty) || arms.len() != 2 || arms.iter().any(|a| a.guard.is_some()) {
-            return None;
-        }
-        // A SCALAR-Ok subject reads len-as-tag (@4); a HEAP-Ok subject is the cap-as-tag
-        // 1-slot block (len@4 always 1, payload handle @12 low-32, tag @16) — the SAME
-        // uniform Err=then(tag≠0)/Ok=else(tag 0) skeleton, only the tag offset and the
-        // Ok-payload load differ (scalar copy vs borrowed LoadHandle — the statement-match
-        // twin's `materialized_results_str` discipline). Opens the desugared
-        // `let v = result.collect_map(..)!; ok(v)` tail (a heap-Ok match returned).
-        let (ok_pay_ty, err_pay_ty) = match &subject.ty {
-            Ty::Applied(TypeConstructorId::Result, a) if a.len() == 2 => {
-                (a[0].clone(), a[1].clone())
-            }
-            _ => return None,
-        };
-        let heap_ok = is_heap_ty(&ok_pay_ty);
-        let tag_off = if heap_ok { 16 } else { 4 };
-        let mut ok_arm: Option<(&IrExpr, Option<VarId>)> = None;
-        let mut err_arm: Option<(&IrExpr, Option<VarId>)> = None;
-        for arm in arms {
-            match &arm.pattern {
-                IrPattern::Ok { inner } => {
-                    let bind = match inner.as_ref() {
-                        IrPattern::Bind { var, .. } => Some(*var),
-                        IrPattern::Wildcard => None,
-                        _ => return None,
-                    };
-                    if ok_arm.is_some() {
-                        return None;
-                    }
-                    ok_arm = Some((&arm.body, bind));
-                }
-                IrPattern::Err { inner } => {
-                    let bind = match inner.as_ref() {
-                        IrPattern::Bind { var, .. } => Some(*var),
-                        IrPattern::Wildcard => None,
-                        _ => return None,
-                    };
-                    if err_arm.is_some() {
-                        return None;
-                    }
-                    err_arm = Some((&arm.body, bind));
-                }
-                IrPattern::Wildcard => {
-                    if err_arm.is_some() {
-                        return None;
-                    }
-                    err_arm = Some((&arm.body, None));
-                }
-                _ => return None,
-            }
-        }
-        let ((ok_body, ok_bind), (err_body, err_bind)) = match (ok_arm, err_arm) {
-            (Some(o), Some(e)) => (o, e),
-            _ => return None,
-        };
-        let ops_mark = self.ops.len();
-        let lifted_mark = self.lifted.len();
-        let lhh_mark = self.live_heap_handles.len();
-        // Materialize the subject to an OWNED tracked temp (a call result is fresh-owned; a
-        // Var is borrowed — Dup it so the scope-end drop discipline is uniform). The temp is
-        // in live_heap_handles → freed by the epilogue AFTER the merge move-out.
-        let subj = match self
-            .lower_call_args(std::slice::from_ref(subject))
-            .ok()
-            .and_then(|a| a.into_iter().next())
-        {
-            Some(CallArg::Handle(v)) => v,
-            _ => {
-                self.ops.truncate(ops_mark);
-                self.lifted.truncate(lifted_mark);
-                self.live_heap_handles.truncate(lhh_mark);
-                return None;
-            }
-        };
-        if self.deferred_opaque_binds.contains(&subj) {
-            self.ops.truncate(ops_mark);
-            self.lifted.truncate(lifted_mark);
-            self.live_heap_handles.truncate(lhh_mark);
-            return None;
-        }
-        if heap_ok {
-            // The cap-as-tag wrapper's epilogue drop must release its @12 payload too:
-            // DropListStr rc_decs the single slot handle (len@4 is always 1) + frees the
-            // wrapper — the Ok arm's Dup keeps a returned payload alive, so the dec is the
-            // borrow's release, never a double-free. (A heap-ERR inner's own nested strings
-            // free flat — the statement twin's `else heap_elem_lists` leak-parity bucket.)
-            self.heap_elem_lists.insert(subj);
-        }
-        let h = self.fresh_value();
-        self.ops.push(Op::Prim { kind: PrimKind::Handle, dst: Some(h), args: vec![subj] });
-        let tag = self.load_at_offset(h, tag_off, PrimKind::Load { width: 4 });
-        let dst = self.fresh_value();
-        self.ops.push(Op::IfThen { cond: tag, dst: Some(dst) });
-        // THEN (tag != 0 = Err): payload = slot 0 (borrowed).
-        if let Some(var) = err_bind {
-            let payload = if is_heap_ty(&err_pay_ty) {
-                let p = self.load_at_offset(h, 12, PrimKind::LoadHandle);
-                self.param_values.insert(p);
-                p
-            } else {
-                self.load_at_offset(h, 12, PrimKind::Load { width: 8 })
-            };
-            self.value_of.insert(var, payload);
-        }
-        let outer: Vec<ValueId> = self.live_heap_handles.clone();
-        let err_obj = match self.lower_heap_result_arm(err_body, result_ty) {
-            Some(v) => v,
-            _ => {
-                self.ops.truncate(ops_mark);
-                self.lifted.truncate(lifted_mark);
-                self.live_heap_handles.truncate(lhh_mark);
-                return None;
-            }
-        };
-        let consumed_by_err: Vec<ValueId> =
-            outer.iter().copied().filter(|x| !self.live_heap_handles.contains(x)).collect();
-        let else_marker_at = self.ops.len();
-        self.ops.push(Op::Else { val: Some(err_obj) });
-        // ELSE (tag == 0 = Ok): a scalar payload copies; a HEAP payload (cap-as-tag
-        // subject) binds the @12 handle as a BORROW — the subject temp still owns it
-        // (freed by the epilogue), an arm that returns it acquires its own ref (Dup).
-        if let Some(var) = ok_bind {
-            let payload = if heap_ok {
-                let p = self.load_at_offset(h, 12, PrimKind::LoadHandle);
-                self.param_values.insert(p);
-                p
-            } else {
-                let _ = &ok_pay_ty;
-                self.load_at_offset(h, 12, PrimKind::Load { width: 8 })
-            };
-            self.value_of.insert(var, payload);
-        }
-        let live_after_err: Vec<ValueId> = self.live_heap_handles.clone();
-        let ok_obj = match self.lower_heap_result_arm(ok_body, result_ty) {
-            Some(v) => v,
-            _ => {
-                self.ops.truncate(ops_mark);
-                self.lifted.truncate(lifted_mark);
-                self.live_heap_handles.truncate(lhh_mark);
-                return None;
-            }
-        };
-        let consumed_by_ok: Vec<ValueId> = live_after_err
-            .iter()
-            .copied()
-            .filter(|x| !self.live_heap_handles.contains(x))
-            .collect();
-        // Release parity across the arms (the lower_heap_result_if_inner discipline).
-        for x in &consumed_by_err {
-            if !consumed_by_ok.contains(x) {
-                let op = self.drop_op_for(*x);
-                self.ops.push(op);
-            }
-        }
-        for x in &consumed_by_ok {
-            if !consumed_by_err.contains(x) {
-                let op = self.drop_op_for(*x);
-                self.ops.insert(else_marker_at, op);
-            }
-        }
-        self.ops.push(Op::EndIf { val: Some(ok_obj) });
-        Some(dst)
-    }
-
-    /// TAIL-VALUE match over a LIST subject with exactly one `[]` arm and one catch-all
-    /// (`_` or a bind-all `ys`) — the len-tag twin of [`Self::try_lower_result_match_value`]:
-    ///   `match list.filter(xs, f) { [] => None, ys => list.get(ys, 0) }`
-    /// The subject is an OWNED tracked temp (a call result is fresh-owned; a Var is Dup'd).
-    /// tag = len@4: THEN (len != 0) = the non-empty arm — a bind-all var ALIASES the subject
-    /// temp itself (arm calls borrow it; if the arm MOVES it out, the release-parity sweep
-    /// compensates with a drop on the empty side). ELSE (len == 0) = the `[]` arm. Same
-    /// IfThen/Else/EndIf merge + release-parity discipline as the Result opener.
-    pub(crate) fn try_lower_list_match_value(
-        &mut self,
-        subject: &IrExpr,
-        arms: &[IrMatchArm],
-        result_ty: &Ty,
-    ) -> Option<ValueId> {
-        use crate::PrimKind;
-        use almide_lang::types::constructor::TypeConstructorId;
-        if !is_heap_ty(result_ty) || arms.len() != 2 || arms.iter().any(|a| a.guard.is_some()) {
-            return None;
-        }
-        if !matches!(&subject.ty, Ty::Applied(TypeConstructorId::List, a) if a.len() == 1) {
-            return None;
-        }
-        let mut empty_arm: Option<&IrExpr> = None;
-        let mut rest_arm: Option<(&IrExpr, Option<VarId>)> = None;
-        for arm in arms {
-            match &arm.pattern {
-                IrPattern::List { elements } if elements.is_empty() => {
-                    if empty_arm.is_some() {
-                        return None;
-                    }
-                    empty_arm = Some(&arm.body);
-                }
-                IrPattern::Bind { var, .. } => {
-                    if rest_arm.is_some() {
-                        return None;
-                    }
-                    rest_arm = Some((&arm.body, Some(*var)));
-                }
-                IrPattern::Wildcard => {
-                    if rest_arm.is_some() {
-                        return None;
-                    }
-                    rest_arm = Some((&arm.body, None));
-                }
-                _ => return None,
-            }
-        }
-        let (empty_body, (rest_body, rest_bind)) = match (empty_arm, rest_arm) {
-            (Some(e), Some(r)) => (e, r),
-            _ => return None,
-        };
-        let ops_mark = self.ops.len();
-        let lifted_mark = self.lifted.len();
-        let lhh_mark = self.live_heap_handles.len();
-        let subj = match self
-            .lower_call_args(std::slice::from_ref(subject))
-            .ok()
-            .and_then(|a| a.into_iter().next())
-        {
-            Some(CallArg::Handle(v)) => v,
-            _ => {
-                self.ops.truncate(ops_mark);
-                self.lifted.truncate(lifted_mark);
-                self.live_heap_handles.truncate(lhh_mark);
-                return None;
-            }
-        };
-        if self.deferred_opaque_binds.contains(&subj) {
-            self.ops.truncate(ops_mark);
-            self.lifted.truncate(lifted_mark);
-            self.live_heap_handles.truncate(lhh_mark);
-            return None;
-        }
-        let h = self.fresh_value();
-        self.ops.push(Op::Prim { kind: PrimKind::Handle, dst: Some(h), args: vec![subj] });
-        let tag = self.load_at_offset(h, 4, PrimKind::Load { width: 4 });
-        let dst = self.fresh_value();
-        self.ops.push(Op::IfThen { cond: tag, dst: Some(dst) });
-        // THEN (len != 0): the non-empty arm; the bind-all aliases the subject temp.
-        if let Some(var) = rest_bind {
-            self.value_of.insert(var, subj);
-        }
-        let outer: Vec<ValueId> = self.live_heap_handles.clone();
-        let rest_obj = match self.lower_heap_result_arm(rest_body, result_ty) {
-            Some(v) => v,
-            _ => {
-                self.ops.truncate(ops_mark);
-                self.lifted.truncate(lifted_mark);
-                self.live_heap_handles.truncate(lhh_mark);
-                return None;
-            }
-        };
-        let consumed_by_rest: Vec<ValueId> =
-            outer.iter().copied().filter(|x| !self.live_heap_handles.contains(x)).collect();
-        let else_marker_at = self.ops.len();
-        self.ops.push(Op::Else { val: Some(rest_obj) });
-        // ELSE (len == 0): the `[]` arm.
-        let live_after_rest: Vec<ValueId> = self.live_heap_handles.clone();
-        let empty_obj = match self.lower_heap_result_arm(empty_body, result_ty) {
-            Some(v) => v,
-            _ => {
-                self.ops.truncate(ops_mark);
-                self.lifted.truncate(lifted_mark);
-                self.live_heap_handles.truncate(lhh_mark);
-                return None;
-            }
-        };
-        let consumed_by_empty: Vec<ValueId> = live_after_rest
-            .iter()
-            .copied()
-            .filter(|x| !self.live_heap_handles.contains(x))
-            .collect();
-        // Release parity across the arms (the lower_heap_result_if_inner discipline).
-        for x in &consumed_by_rest {
-            if !consumed_by_empty.contains(x) {
-                let op = self.drop_op_for(*x);
-                self.ops.push(op);
-            }
-        }
-        for x in &consumed_by_empty {
-            if !consumed_by_rest.contains(x) {
-                let op = self.drop_op_for(*x);
-                self.ops.insert(else_marker_at, op);
-            }
-        }
-        self.ops.push(Op::EndIf { val: Some(empty_obj) });
-        Some(dst)
-    }
-
-    /// Is `pat` structurally valid for the NESTED-refinement chain over variant
-    /// `tyname` — every ctor resolves in the layout with matching arity, every
-    /// nested arg a wildcard / Int-Bool literal (over a scalar field) / a ctor
-    /// over a variant-typed field (recursively)? A Bind or any other shape → false
-    /// (the chain has no arm-scope payload binds).
-    fn nested_refinement_pat_valid(&self, pat: &IrPattern, tyname: &str) -> bool {
-        let IrPattern::Constructor { name, args } = pat else { return false };
-        let Some(layout) = self.variant_layouts.by_type.get(tyname) else { return false };
-        let Some(case) = layout.cases.iter().find(|c| c.ctor.as_str() == name.as_str()) else {
-            return false;
-        };
-        if args.len() != case.fields.len() && !args.is_empty() {
-            return false;
-        }
-        args.iter().zip(case.fields.iter()).all(|(a, (_, fty))| match a {
-            IrPattern::Wildcard => true,
-            IrPattern::Literal { expr } => {
-                !is_heap_ty(fty)
-                    && matches!(expr.kind, IrExprKind::LitInt { .. } | IrExprKind::LitBool { .. })
-            }
-            IrPattern::Constructor { .. } => self
-                .custom_variant_type_name(fty)
-                .is_some_and(|inner| self.nested_refinement_pat_valid(a, &inner)),
-            _ => false,
-        })
-    }
-
-    /// Emit the SHORT-CIRCUIT refinement condition for `pat` over the (borrowed)
-    /// variant block `block` of type `tyname`: outer tag equality, and — ONLY when
-    /// it holds (a FLAT-MARKER if, so an out-of-case slot is never dereferenced) —
-    /// the conjunction (0/1 multiply) of every nested tag/literal refinement.
-    /// All reads are borrows (Handle/Load/LoadHandle — no ownership event).
-    /// Caller guarantees [`Self::nested_refinement_pat_valid`].
-    fn nested_refinement_cond(
-        &mut self,
-        block: ValueId,
-        pat: &IrPattern,
-        tyname: &str,
-    ) -> Option<ValueId> {
-        use crate::PrimKind;
-        let IrPattern::Constructor { name, args } = pat else { return None };
-        let layout = self.variant_layouts.by_type.get(tyname)?.clone();
-        let case = layout.cases.iter().find(|c| c.ctor.as_str() == name.as_str())?.clone();
-        let h = self.fresh_value();
-        self.ops.push(Op::Prim { kind: PrimKind::Handle, dst: Some(h), args: vec![block] });
-        let tagv = self.load_at_offset(h, 12, PrimKind::Load { width: 8 });
-        let want = self.fresh_value();
-        self.ops.push(Op::ConstInt { dst: want, value: case.tag as i64 });
-        let eq = self.fresh_value();
-        self.ops.push(Op::IntBinOp { dst: eq, op: IntOp::Eq, a: tagv, b: want });
-        let refut: Vec<(usize, &IrPattern)> = args
-            .iter()
-            .enumerate()
-            .filter(|(_, a)| matches!(a, IrPattern::Constructor { .. } | IrPattern::Literal { .. }))
-            .collect();
-        if refut.is_empty() {
-            return Some(eq);
-        }
-        let dst = self.fresh_value();
-        self.ops.push(Op::IfThen { cond: eq, dst: Some(dst) });
-        let mut conj: Option<ValueId> = None;
-        for (i, sub) in refut {
-            let off = (12 + 8 * (i + 1)) as i64;
-            let ci = match sub {
-                IrPattern::Literal { expr } => {
-                    let fv = self.load_at_offset(h, off, PrimKind::Load { width: 8 });
-                    let lit = self.fresh_value();
-                    let value = match &expr.kind {
-                        IrExprKind::LitInt { value } => *value,
-                        IrExprKind::LitBool { value } => *value as i64,
-                        _ => return None,
-                    };
-                    self.ops.push(Op::ConstInt { dst: lit, value });
-                    let c = self.fresh_value();
-                    self.ops.push(Op::IntBinOp { dst: c, op: IntOp::Eq, a: fv, b: lit });
-                    c
-                }
-                IrPattern::Constructor { .. } => {
-                    let fblock = self.load_at_offset(h, off, PrimKind::LoadHandle);
-                    let inner = self.custom_variant_type_name(&case.fields[i].1)?;
-                    self.nested_refinement_cond(fblock, sub, &inner)?
-                }
-                _ => unreachable!("filtered above"),
-            };
-            conj = Some(match conj {
-                None => ci,
-                Some(prev) => {
-                    let c = self.fresh_value();
-                    self.ops.push(Op::IntBinOp { dst: c, op: IntOp::Mul, a: prev, b: ci });
-                    c
-                }
-            });
-        }
-        self.ops.push(Op::Else { val: Some(conj?) });
-        let zero = self.fresh_value();
-        self.ops.push(Op::ConstInt { dst: zero, value: 0 });
-        self.ops.push(Op::EndIf { val: Some(zero) });
-        Some(dst)
-    }
-
-    /// The ordered arm chain of the nested-refinement match: `if cond1 then body1
-    /// else if cond2 then body2 … else <irrefutable body>` over the flat markers.
-    fn nested_refinement_chain(
-        &mut self,
-        subj: ValueId,
-        arms: &[IrMatchArm],
-        tyname: &str,
-    ) -> Option<ValueId> {
-        let arm = arms.first()?;
-        if matches!(arm.pattern, IrPattern::Wildcard) {
-            return self.lower_scalar_arm(&arm.body);
-        }
-        let cond = self.nested_refinement_cond(subj, &arm.pattern, tyname)?;
-        let dst = self.fresh_value();
-        self.ops.push(Op::IfThen { cond, dst: Some(dst) });
-        let then_v = self.lower_scalar_arm(&arm.body)?;
-        self.ops.push(Op::Else { val: Some(then_v) });
-        let rest_v = self.nested_refinement_chain(subj, &arms[1..], tyname)?;
-        self.ops.push(Op::EndIf { val: Some(rest_v) });
-        Some(dst)
-    }
-
-    /// EXECUTE a `match` over a custom variant with NESTED refutable constructor
-    /// patterns (`Node(Red, Node(_, _, 5, _), _, _)` — the C-070 boxed-refinement
-    /// class) as an ORDERED refinement chain: each arm's condition is the outer tag
-    /// equality plus every inner tag/literal constraint, evaluated SHORT-CIRCUIT
-    /// (inner slots read only under their ctor's matching tag — no out-of-case
-    /// dereference), then dispatched first-match-wins via the flat-marker if chain.
-    /// ADMITTED: ctor/wildcard/Int-Bool-literal args (a BINDER would need arm-scope
-    /// payload binds — declined), a SCALAR result, no guards, a trailing wildcard,
-    /// and a Var subject that is a real block (param / materialized). Every payload
-    /// read is a BORROW; the whole attempt rolls back on any miss.
-    fn try_lower_nested_refinement_match(
-        &mut self,
-        subject: &IrExpr,
-        arms: &[IrMatchArm],
-        result_ty: &Ty,
-    ) -> Option<ValueId> {
-        if is_heap_ty(result_ty) {
-            return None;
-        }
-        let type_name = self.custom_variant_type_name(&subject.ty)?;
-        // Only fire when some arm actually NESTS a refutable arg — the flat
-        // tag-switch machinery owns every other shape.
-        let has_nested = |p: &IrPattern| {
-            matches!(p, IrPattern::Constructor { args, .. }
-                if args.iter().any(|a| matches!(a,
-                    IrPattern::Constructor { .. } | IrPattern::Literal { .. })))
-        };
-        if !arms.iter().any(|a| has_nested(&a.pattern)) {
-            return None;
-        }
-        if !matches!(arms.last()?.pattern, IrPattern::Wildcard) {
-            return None;
-        }
-        if !arms[..arms.len() - 1]
-            .iter()
-            .all(|a| self.nested_refinement_pat_valid(&a.pattern, &type_name))
-        {
-            return None;
-        }
-        // The subject must be a REAL block: a borrowed variant param or a locally
-        // materialized value (a deferred Opaque's tag read would be garbage).
-        let IrExprKind::Var { id } = &subject.kind else { return None };
-        let subj = self.value_for(*id).ok()?;
-        if !self.param_values.contains(&subj) && !self.materialized_aggregates.contains(&subj) {
-            return None;
-        }
-        let ops_mark = self.ops.len();
-        let lhh_mark = self.live_heap_handles.len();
-        match self.nested_refinement_chain(subj, arms, &type_name) {
-            Some(dst) => Some(dst),
-            None => {
-                if std::env::var("ALMIDE_DBG_NESTED_MATCH").is_ok() {
-                    eprintln!("[nested-match] chain declined for {}", self.fn_name);
-                }
-                self.ops.truncate(ops_mark);
-                self.live_heap_handles.truncate(lhh_mark);
-                None
-            }
-        }
-    }
-
-    /// EXECUTE a `match (a, b, …) { (P1, P2, …) => …, _ => … }` over a TUPLE of
-    /// BOUND VARS — the frontend's factored form of a multi-arm nested-ctor match
-    /// (`match t { Node(Red, Node(...), _, _) => …, Node(...) => …, … }` desugars
-    /// to ONE outer `Node(a,b,c,d)` arm whose body re-matches `(a,b,c,d)` — the
-    /// C-070 class). Each arm's condition is the CONJUNCTION of per-element
-    /// refinements (a variant element via [`Self::nested_refinement_cond`], a
-    /// scalar element via literal equality; wildcards free), dispatched
-    /// first-match-wins on the flat-marker chain. Scalar result, no guards, a
-    /// trailing wildcard, all elements plain Vars.
-    pub(crate) fn try_lower_tuple_refinement_match(
-        &mut self,
-        subject: &IrExpr,
-        arms: &[IrMatchArm],
-        result_ty: &Ty,
-    ) -> Option<ValueId> {
-        let dbg = std::env::var("ALMIDE_DBG_NESTED_MATCH").is_ok();
-        if dbg {
-            eprintln!("[tuple-match] entry: subj={:?} arms={}", std::mem::discriminant(&subject.kind), arms.len());
-        }
-        if arms.is_empty() || arms.iter().any(|a| a.guard.is_some()) {
-            return None;
-        }
-        let IrExprKind::Tuple { elements } = &subject.kind else { return None };
-        if !matches!(arms.last()?.pattern, IrPattern::Wildcard) {
-            if dbg {
-                eprintln!("[tuple-match] last arm not wildcard");
-            }
-            return None;
-        }
-        // The marks precede the ELEMENT resolution: a scalar-EXPRESSION element
-        // (`n % 3` — the fizz shape) ANF-lowers here, so every decline below must
-        // roll its ops back.
-        let ops_mark = self.ops.len();
-        let lhh_mark = self.live_heap_handles.len();
-        let mut rollback = |s: &mut Self| {
-            s.ops.truncate(ops_mark);
-            s.live_heap_handles.truncate(lhh_mark);
-        };
-        // Elements: a plain Var (variant block or scalar), or a SCALAR expression
-        // materialized to a fresh value (read once — exactly the by-value tuple
-        // subject semantics; the match reads only these copies).
-        let mut elems: Vec<(ValueId, Ty)> = Vec::with_capacity(elements.len());
-        for e in elements {
-            let v = match &e.kind {
-                IrExprKind::Var { id } => match self.value_for(*id) {
-                    Ok(v) => v,
-                    Err(_) => {
-                        rollback(self);
-                        return None;
-                    }
-                },
-                _ if !is_heap_ty(&e.ty) => match self.lower_scalar_value(e) {
-                    Some(v) => v,
-                    None => {
-                        rollback(self);
-                        return None;
-                    }
-                },
-                _ => {
-                    rollback(self);
-                    return None;
-                }
-            };
-            elems.push((v, e.ty.clone()));
-        }
-        // Validate every refutable arm up front (no mid-emission decline).
-        for a in &arms[..arms.len() - 1] {
-            let IrPattern::Tuple { elements: pats } = &a.pattern else {
-                rollback(self);
-                return None;
-            };
-            if pats.len() != elems.len() {
-                rollback(self);
-                return None;
-            }
-            for (p, (_, ty)) in pats.iter().zip(elems.iter()) {
-                let ok = match p {
-                    IrPattern::Wildcard => true,
-                    IrPattern::Literal { expr } => {
-                        !is_heap_ty(ty)
-                            && matches!(
-                                expr.kind,
-                                IrExprKind::LitInt { .. } | IrExprKind::LitBool { .. }
-                            )
-                    }
-                    IrPattern::Constructor { .. } => self
-                        .custom_variant_type_name(ty)
-                        .is_some_and(|n| self.nested_refinement_pat_valid(p, &n)),
-                    _ => false,
-                };
-                if !ok {
-                    rollback(self);
-                    return None;
-                }
-            }
-        }
-        match self.tuple_refinement_chain(&elems, arms, result_ty) {
-            Some(dst) => Some(dst),
-            None => {
-                self.ops.truncate(ops_mark);
-                self.live_heap_handles.truncate(lhh_mark);
-                None
-            }
-        }
-    }
-
-    /// The UNIT-statement sibling of [`Self::try_lower_tuple_refinement_match`] —
-    /// the heap-branch tail duplication turns `let s = match (a, b) { … }; use(s)`
-    /// into a STATEMENT match whose arms carry the continuation's effects, so the
-    /// refinement chain must run them under REAL `IfThen`/`Else`/`EndIf` markers
-    /// (only the taken arm executes — `unit_arm_depth` raised per arm, exactly the
-    /// `lower_variant_unit_arm` discipline). Returns `true` iff fully lowered;
-    /// rolls back and returns `false` on any decline.
-    pub(crate) fn try_lower_tuple_refinement_unit_match(
-        &mut self,
-        subject: &IrExpr,
-        arms: &[IrMatchArm],
-    ) -> bool {
-        if arms.is_empty() || arms.iter().any(|a| a.guard.is_some()) {
-            return false;
-        }
-        let IrExprKind::Tuple { elements } = &subject.kind else { return false };
-        if !matches!(arms.last().map(|a| &a.pattern), Some(IrPattern::Wildcard)) {
-            return false;
-        }
-        let ops_mark = self.ops.len();
-        let lhh_mark = self.live_heap_handles.len();
-        let mut elems: Vec<(ValueId, Ty)> = Vec::with_capacity(elements.len());
-        for e in elements {
-            let v = match &e.kind {
-                IrExprKind::Var { id } => self.value_for(*id).ok(),
-                _ if !is_heap_ty(&e.ty) => self.lower_scalar_value(e),
-                _ => None,
-            };
-            match v {
-                Some(v) => elems.push((v, e.ty.clone())),
-                None => {
-                    self.ops.truncate(ops_mark);
-                    self.live_heap_handles.truncate(lhh_mark);
-                    return false;
-                }
-            }
-        }
-        let mut valid = true;
-        'outer: for a in &arms[..arms.len() - 1] {
-            let IrPattern::Tuple { elements: pats } = &a.pattern else {
-                valid = false;
-                break;
-            };
-            if pats.len() != elems.len() {
-                valid = false;
-                break;
-            }
-            for (p, (_, ty)) in pats.iter().zip(elems.iter()) {
-                let ok = match p {
-                    IrPattern::Wildcard => true,
-                    IrPattern::Literal { expr } => {
-                        !is_heap_ty(ty)
-                            && matches!(
-                                expr.kind,
-                                IrExprKind::LitInt { .. } | IrExprKind::LitBool { .. }
-                            )
-                    }
-                    IrPattern::Constructor { .. } => self
-                        .custom_variant_type_name(ty)
-                        .is_some_and(|n| self.nested_refinement_pat_valid(p, &n)),
-                    _ => false,
-                };
-                if !ok {
-                    valid = false;
-                    break 'outer;
-                }
-            }
-        }
-        if !valid || !self.tuple_refinement_unit_chain(&elems, arms) {
-            self.ops.truncate(ops_mark);
-            self.live_heap_handles.truncate(lhh_mark);
-            return false;
-        }
-        true
-    }
-
-    fn tuple_refinement_unit_chain(
-        &mut self,
-        elems: &[(ValueId, Ty)],
-        arms: &[IrMatchArm],
-    ) -> bool {
-        let Some(arm) = arms.first() else { return true };
-        let lower_unit_arm = |s: &mut Self, body: &IrExpr| -> bool {
-            let mark = s.live_heap_handles.len();
-            s.unit_arm_depth += 1;
-            let r = s.lower_branch_arm(None, body);
-            s.unit_arm_depth -= 1;
-            if r.is_err() {
-                return false;
-            }
-            s.drop_arm_locals(mark);
-            true
-        };
-        if matches!(arm.pattern, IrPattern::Wildcard) {
-            return lower_unit_arm(self, &arm.body);
-        }
-        let IrPattern::Tuple { elements: pats } = &arm.pattern else { return false };
-        let mut conj: Option<ValueId> = None;
-        for (p, (v, ty)) in pats.iter().zip(elems.iter()) {
-            let ci = match p {
-                IrPattern::Wildcard => continue,
-                IrPattern::Literal { expr } => {
-                    let lit = self.fresh_value();
-                    let value = match &expr.kind {
-                        IrExprKind::LitInt { value } => *value,
-                        IrExprKind::LitBool { value } => *value as i64,
-                        _ => return false,
-                    };
-                    self.ops.push(Op::ConstInt { dst: lit, value });
-                    let c = self.fresh_value();
-                    self.ops.push(Op::IntBinOp { dst: c, op: IntOp::Eq, a: *v, b: lit });
-                    c
-                }
-                IrPattern::Constructor { .. } => {
-                    let Some(vname) = self.custom_variant_type_name(ty) else { return false };
-                    match self.nested_refinement_cond(*v, p, &vname) {
-                        Some(c) => c,
-                        None => return false,
-                    }
-                }
-                _ => return false,
-            };
-            conj = Some(match conj {
-                None => ci,
-                Some(prev) => {
-                    let c = self.fresh_value();
-                    self.ops.push(Op::IntBinOp { dst: c, op: IntOp::Mul, a: prev, b: ci });
-                    c
-                }
-            });
-        }
-        let Some(cond) = conj else {
-            return lower_unit_arm(self, &arm.body);
-        };
-        self.ops.push(Op::IfThen { cond, dst: None });
-        if !lower_unit_arm(self, &arm.body) {
-            return false;
-        }
-        self.ops.push(Op::Else { val: None });
-        if !self.tuple_refinement_unit_chain(elems, &arms[1..]) {
-            return false;
-        }
-        self.ops.push(Op::EndIf { val: None });
-        true
-    }
-
-    fn tuple_refinement_chain(
-        &mut self,
-        elems: &[(ValueId, Ty)],
-        arms: &[IrMatchArm],
-        result_ty: &Ty,
-    ) -> Option<ValueId> {
-        let heap = is_heap_ty(result_ty);
-        let lower_arm = |s: &mut Self, body: &IrExpr| -> Option<ValueId> {
-            if heap {
-                s.lower_heap_result_arm(body, result_ty)
-            } else {
-                s.lower_scalar_arm(body)
-            }
-        };
-        let arm = arms.first()?;
-        if matches!(arm.pattern, IrPattern::Wildcard) {
-            return lower_arm(self, &arm.body);
-        }
-        let IrPattern::Tuple { elements: pats } = &arm.pattern else { return None };
-        let mut conj: Option<ValueId> = None;
-        for (p, (v, ty)) in pats.iter().zip(elems.iter()) {
-            let ci = match p {
-                IrPattern::Wildcard => continue,
-                IrPattern::Literal { expr } => {
-                    let lit = self.fresh_value();
-                    let value = match &expr.kind {
-                        IrExprKind::LitInt { value } => *value,
-                        IrExprKind::LitBool { value } => *value as i64,
-                        _ => return None,
-                    };
-                    self.ops.push(Op::ConstInt { dst: lit, value });
-                    let c = self.fresh_value();
-                    self.ops.push(Op::IntBinOp { dst: c, op: IntOp::Eq, a: *v, b: lit });
-                    c
-                }
-                IrPattern::Constructor { .. } => {
-                    let vname = self.custom_variant_type_name(ty)?;
-                    self.nested_refinement_cond(*v, p, &vname)?
-                }
-                _ => return None,
-            };
-            conj = Some(match conj {
-                None => ci,
-                Some(prev) => {
-                    let c = self.fresh_value();
-                    self.ops.push(Op::IntBinOp { dst: c, op: IntOp::Mul, a: prev, b: ci });
-                    c
-                }
-            });
-        }
-        // An all-wildcard tuple arm is irrefutable — its body IS the rest.
-        let Some(cond) = conj else {
-            return lower_arm(self, &arm.body);
-        };
-        let dst = self.fresh_value();
-        self.ops.push(Op::IfThen { cond, dst: Some(dst) });
-        // RELEASE PARITY for the HEAP merge (mirrors emit_variant_arm_chain): an
-        // outer handle one side moves out must be released by the other, else the
-        // branch-grouped cert rejects the path-dependent accounting. The scalar
-        // path emits no ownership events, so the parity sets stay empty there.
-        let outer: Vec<ValueId> = self.live_heap_handles.clone();
-        let then_v = lower_arm(self, &arm.body)?;
-        let consumed_by_then: Vec<ValueId> =
-            outer.iter().copied().filter(|x| !self.live_heap_handles.contains(x)).collect();
-        let else_marker_at = self.ops.len();
-        self.ops.push(Op::Else { val: Some(then_v) });
-        let live_after_then: Vec<ValueId> = self.live_heap_handles.clone();
-        let rest_v = self.tuple_refinement_chain(elems, &arms[1..], result_ty)?;
-        let consumed_by_else: Vec<ValueId> = live_after_then
-            .iter()
-            .copied()
-            .filter(|x| !self.live_heap_handles.contains(x))
-            .collect();
-        for x in &consumed_by_then {
-            if !consumed_by_else.contains(x) {
-                let op = self.drop_op_for(*x);
-                self.ops.push(op);
-            }
-        }
-        for x in &consumed_by_else {
-            if !consumed_by_then.contains(x) {
-                let op = self.drop_op_for(*x);
-                self.ops.insert(else_marker_at, op);
-            }
-        }
-        self.ops.push(Op::EndIf { val: Some(rest_v) });
-        Some(dst)
-    }
-
-    pub(crate) fn try_lower_custom_variant_match(
-        &mut self,
-        subject: &IrExpr,
-        arms: &[IrMatchArm],
-        result_ty: &Ty,
-    ) -> Option<ValueId> {
-        use crate::PrimKind;
-        if arms.is_empty() || arms.iter().any(|a| a.guard.is_some()) {
-            return None;
-        }
-        // C-070: nested refutable ctor/literal patterns — the ordered refinement
-        // chain (declines untouched shapes; the tag-switch below owns the rest).
-        if let Some(dst) = self.try_lower_nested_refinement_match(subject, arms, result_ty) {
-            return Some(dst);
-        }
-        // The subject must be a registered custom variant; clone its layout out of the borrow.
-        let type_name = self.custom_variant_type_name(&subject.ty)?;
-        let layout = self.variant_layouts.by_type.get(&type_name)?.clone();
-        // DEPTH-2 patterns over a SINGLE-CTOR outer (`match o { Wrap(A(n)) => …, Wrap(B(m))
-        // => … }` — the `pick` shape): the one outer ctor ALWAYS matches, so the match IS the
-        // inner dispatch over the payload. STRIP the outer layer (arms' patterns become the
-        // inner patterns; the inner layout comes from the payload field's variant type) and
-        // remember to UNWRAP one level: the dispatch handle becomes the payload's slot-1
-        // handle (a BORROW — the subject keeps owning it; loaded only after the subject
-        // materializes, so no wrong-ctor garbage read is possible: there IS no other ctor).
-        let mut layout = layout;
-        let mut stripped: Vec<IrMatchArm>;
-        let mut arms: &[IrMatchArm] = arms;
-        let mut unwrap_single = false;
-        // Guard-clause flattening of the former 6-deep nested-if (no `else` anywhere: any
-        // unmet condition falls through to the code after this block, unchanged, exactly as
-        // the original fell out of the if-pyramid — `break` exits the labeled block and
-        // resumes there). `case`'s borrow of the OLD `layout` still ends before `layout =
-        // inner_layout` reassigns it (its last use, inside `all_nested`, is unchanged and
-        // stays textually before the reassignment). No behavior change — see
-        // docs/roadmap/active/code-health-codopsy.md.
-        'single_ctor_strip: {
-            if layout.cases.len() != 1 {
-                break 'single_ctor_strip;
-            }
-            let case = &layout.cases[0];
-            if case.fields.len() != 1 {
-                break 'single_ctor_strip;
-            }
-            let all_nested = arms.iter().all(|a| matches!(&a.pattern,
-                IrPattern::Constructor { name, args }
-                    if *name == case.ctor && args.len() == 1
-                        && matches!(args[0], IrPattern::Constructor { .. })));
-            if !all_nested {
-                break 'single_ctor_strip;
-            }
-            let inner_ty = case.fields[0].1.clone();
-            let Some(inner_name) = self.custom_variant_type_name(&inner_ty) else {
-                break 'single_ctor_strip;
-            };
-            let Some(inner_layout) = self.variant_layouts.by_type.get(&inner_name).cloned()
-            else {
-                break 'single_ctor_strip;
-            };
-            stripped = Vec::with_capacity(arms.len());
-            for a in arms {
-                let IrPattern::Constructor { args, .. } = &a.pattern else {
-                    unreachable!("gated above")
-                };
-                stripped.push(IrMatchArm {
-                    pattern: args[0].clone(),
-                    guard: a.guard.clone(),
-                    body: a.body.clone(),
-                });
-            }
-            layout = inner_layout;
-            arms = &stripped;
-            unwrap_single = true;
-        }
-        let plans = self.parse_variant_arms(&layout, arms)?;
-        // A SINGLE-arm HEAP-result match (a 1-ctor newtype `unbox`, `match b { B(x) => x }`) that
-        // returned the arm value DIRECTLY to `func.ret` would double-move (the arm's move-out
-        // `Consume` + the ret's move — the `amm`/`aamdm` net-−1 the proven checker REJECTS). A
-        // 1-CTOR variant's tag ALWAYS matches (there is no other constructor), so route the arm
-        // through an IfThen `dst` (one ret move, exactly like a multi-arm match) whose ELSE is an
-        // unreachable empty-heap block — never executed, so no leak. A single-arm WILDCARD (`_ =>
-        // body`) has no ctor tag to test → stays declined (a later brick). See
-        // [`Self::emit_single_ctor_heap_arm`].
-        let sole_ctor_heap = is_heap_ty(result_ty)
-            && plans.len() == 1
-            && matches!(plans[0].0, VariantArmKind::Ctor { .. });
-        if is_heap_ty(result_ty) && plans.len() == 1 && !sole_ctor_heap {
-            return None;
-        }
-        let ops_mark = self.ops.len();
-        let lifted_mark = self.lifted.len();
-        let lhh_mark = self.live_heap_handles.len();
-        // Materialize/borrow the subject → a Handle (the variant block pointer).
-        let subj = match self
-            .lower_call_args(std::slice::from_ref(subject))
-            .ok()
-            .and_then(|a| a.into_iter().next())
-        {
-            Some(CallArg::Handle(v)) => v,
-            _ => {
-                self.ops.truncate(ops_mark);
-                self.lifted.truncate(lifted_mark);
-                self.live_heap_handles.truncate(lhh_mark);
-                return None;
-            }
-        };
-        // A DEFERRED-Opaque subject is an EMPTY block: reading its tag would take a wrong
-        // arm silently (the record-ctor mt2 miscompile) — decline (the tail walls honestly).
-        if self.deferred_opaque_binds.contains(&subj) {
-            self.ops.truncate(ops_mark);
-            self.lifted.truncate(lifted_mark);
-            self.live_heap_handles.truncate(lhh_mark);
-            return None;
-        }
-        // A HEAP result over an OWNED subject temp would overlap the owned-subject borrow with the
-        // arm's heap move-out (the cert rejects it). Subject-drop-before-arms is ADT brick 4b —
-        // for now WALL it (a borrowed param/var subject, the recursive-to_string case, proceeds).
-        if is_heap_ty(result_ty) && self.live_heap_handles.contains(&subj) {
-            self.ops.truncate(ops_mark);
-            self.lifted.truncate(lifted_mark);
-            self.live_heap_handles.truncate(lhh_mark);
-            return None;
-        }
-        // Read the tag from slot 0, then emit the per-arm if-chain.
-        let h = self.fresh_value();
-        self.ops.push(Op::Prim { kind: PrimKind::Handle, dst: Some(h), args: vec![subj] });
-        // The depth-2 single-outer unwrap: the dispatch handle becomes the payload's
-        // slot-1 handle (BORROWED — the subject owns it; freed by the subject's own
-        // recursive drop, so param_values keeps it un-dropped here).
-        let h = if unwrap_single {
-            let payload = self.load_at_offset(
-                h,
-                layout::slot_offset(1) as i64,
-                PrimKind::LoadHandle,
-            );
-            self.param_values.insert(payload);
-            let ph = self.fresh_value();
-            self.ops.push(Op::Prim { kind: PrimKind::Handle, dst: Some(ph), args: vec![payload] });
-            ph
-        } else {
-            h
-        };
-        // Rung-5 variants slab: the tag is slot 0 of the subject block — the
-        // TARGET-NEUTRAL `ListGetScalar` reads it on both legs. The depth-2
-        // unwrap (`h` re-pointed at the payload block) keeps the raw load: its
-        // container is the borrowed payload, not `subj`.
-        let tag = if unwrap_single {
-            self.load_at_offset(h, layout::slot_offset(0) as i64, PrimKind::Load { width: 8 })
-        } else {
-            let idx = self.fresh_value();
-            self.ops.push(Op::ConstInt { dst: idx, value: 0 });
-            let t = self.fresh_value();
-            self.ops.push(Op::ListGetScalar { dst: t, list: subj, idx });
-            t
-        };
-        let emitted = if sole_ctor_heap {
-            let (kind, body) = &plans[0];
-            self.emit_single_ctor_heap_arm(h, tag, kind, body, result_ty, subj)
-        } else {
-            self.emit_variant_arm_chain(h, tag, &plans, result_ty, subj)
-        };
-        match emitted {
-            Some(dst) => Some(dst),
-            None => {
-                self.ops.truncate(ops_mark);
-                self.lifted.truncate(lifted_mark);
-                self.live_heap_handles.truncate(lhh_mark);
-                None
-            }
-        }
-    }
-
-    /// Route a SOLE-constructor HEAP-result arm through an IfThen `dst` (one ret move) with an
-    /// unreachable empty-heap ELSE. A 1-ctor variant's tag always equals `arm_tag`, so the `else` is
-    /// dead — it exists only so the arm value flows through the branch-merge `dst` the ownership
-    /// certificate needs (a direct return would double-move — see the caller). The empty-heap block is
-    /// never allocated at runtime, so it cannot leak.
-    fn emit_single_ctor_heap_arm(
-        &mut self,
-        h: ValueId,
-        tag: ValueId,
-        kind: &VariantArmKind,
-        body: &IrExpr,
-        result_ty: &Ty,
-        subj: ValueId,
-    ) -> Option<ValueId> {
-        let arm_tag = match kind {
-            VariantArmKind::Ctor { tag, .. } => *tag,
-            VariantArmKind::Wildcard | VariantArmKind::BindAll { .. } => return None,
-        };
-        let dst = self.fresh_value();
-        let tc = self.fresh_value();
-        self.ops.push(Op::ConstInt { dst: tc, value: arm_tag });
-        let cond = self.fresh_value();
-        self.ops.push(Op::IntBinOp { dst: cond, op: IntOp::Eq, a: tag, b: tc });
-        self.ops.push(Op::IfThen { cond, dst: Some(dst) });
-        let then_v = self.lower_variant_arm_value(kind, body, h, result_ty, true, subj)?;
-        self.ops.push(Op::Else { val: Some(then_v) });
-        // The dead `else`: a fresh owned empty-string block (a heap i32 handle, repr-compatible with
-        // any heap result). `Consume` moves it into `dst` exactly as a real arm value would.
-        let repr = repr_of(result_ty).ok()?;
-        let else_v = self.fresh_value();
-        self.ops.push(Op::Alloc { dst: else_v, repr, init: crate::Init::Str(String::new()) });
-        self.ops.push(Op::Consume { v: else_v });
-        self.ops.push(Op::EndIf { val: Some(else_v) });
-        Some(dst)
-    }
-
-    /// Parse a custom-variant `match`'s arms into per-arm plans — shared by the value-result
-    /// ([`Self::try_lower_custom_variant_match`]) and Unit-statement
-    /// ([`Self::lower_custom_variant_unit_match`]) paths. `None` (the caller walls / declines)
-    /// if any arm is outside the scalar-field subset: a guard, a heap-field bind, a nested ctor
-    /// pattern, or a binder catch-all `x => …` (all later bricks). The bodies stay borrowed
-    /// from `arms` (a param, not `self`) — no borrow conflict with the lowering that follows.
-    fn parse_variant_arms<'a>(
-        &self,
-        layout: &VariantLayout,
-        arms: &'a [IrMatchArm],
-    ) -> Option<Vec<(VariantArmKind, &'a IrExpr)>> {
-        let mut plans: Vec<(VariantArmKind, &IrExpr)> = Vec::with_capacity(arms.len());
-        for arm in arms {
-            if arm.guard.is_some() {
-                return None;
-            }
-            let kind = match &arm.pattern {
-                IrPattern::Constructor { name, args } => {
-                    let case = layout.case_by_ctor(name)?;
-                    if args.len() != case.fields.len() {
-                        return None;
-                    }
-                    let mut binds = Vec::new();
-                    for (i, fp) in args.iter().enumerate() {
-                        match fp {
-                            IrPattern::Wildcard => {}
-                            // slot 1+i (slot 0 is the tag). A SCALAR field binds by value copy.
-                            IrPattern::Bind { var, ty } if !is_heap_ty(ty) => {
-                                binds.push((1 + i, *var, false, ty.clone()))
-                            }
-                            // ANY heap field (`String`, a nested VARIANT, a `List[…]` —
-                            // `ArrV(xs) => for x in xs`, the gguf ValArray consumer — Bytes,
-                            // Matrix) binds as a BORROW of the slot handle: the subject owns
-                            // it, a move-out auto-Dups, a borrow-pass just reads. The bind is
-                            // type-agnostic (a slot-handle load); what the ARM does with it is
-                            // gated by the arm-body lowering as usual.
-                            IrPattern::Bind { var, ty } if is_heap_ty(ty) => {
-                                binds.push((1 + i, *var, true, ty.clone()))
-                            }
-                            // a nested ctor pattern — a later brick.
-                            _ => return None,
-                        }
-                    }
-                    VariantArmKind::Ctor { tag: case.tag as i64, binds }
-                }
-                // A RECORD-ctor pattern (`Node { left, right, value }`, `Data { seq, .. }`,
-                // `Click { .. }`): resolve each named field to its declared slot (1 + index)
-                // and bind exactly like the positional ctor arm — scalar by value copy, heap
-                // as a borrow of the slot handle. `..`/unmentioned fields bind nothing; a
-                // NESTED field pattern stays a later brick.
-                IrPattern::RecordPattern { name, fields, rest: _ } => {
-                    let case = layout.case_by_ctor(name)?;
-                    let mut binds = Vec::new();
-                    for f in fields {
-                        let idx = case
-                            .fields
-                            .iter()
-                            .position(|(n, _)| n.as_str() == f.name)?;
-                        match &f.pattern {
-                            None | Some(IrPattern::Wildcard) => {}
-                            Some(IrPattern::Bind { var, ty }) if !is_heap_ty(ty) => {
-                                binds.push((1 + idx, *var, false, ty.clone()))
-                            }
-                            Some(IrPattern::Bind { var, ty }) if is_heap_ty(ty) => {
-                                binds.push((1 + idx, *var, true, ty.clone()))
-                            }
-                            _ => return None,
-                        }
-                    }
-                    VariantArmKind::Ctor { tag: case.tag as i64, binds }
-                }
-                IrPattern::Wildcard => VariantArmKind::Wildcard,
-                // A BINDER catch-all (`e => …`): binds the whole subject (borrow), any tag.
-                IrPattern::Bind { var, ty } if is_heap_ty(ty) => {
-                    VariantArmKind::BindAll { var: *var }
-                }
-                _ => return None,
-            };
-            plans.push((kind, &arm.body));
-        }
-        Some(plans)
-    }
-
-    /// Bind a custom-variant arm's ctor fields from the block's slots (a `Wildcard` arm binds
-    /// nothing). A SCALAR field is an i64 value COPY (`Load`); a leaf-heap (`String`) field is a
-    /// `Dup`'d OWNED copy of the slot handle (`LoadHandle` then `Op::Dup`, rc+1) pushed to
-    /// `live_heap_handles` so the ARM FRAME drops it at arm end (`emit_variant_arm_chain` marks
-    /// before this call). The OWNED copy — not a borrow — is what the proven checker needs: a
-    /// consuming re-use moves an owned ref, a read-only use drops it, a move-out hands it off,
-    /// all rc-balanced; a BORROW would `Consume`/`m` at rc 0 on a re-use (the rejected double-free).
-    fn bind_variant_arm(&mut self, kind: &VariantArmKind, h: ValueId, subj: ValueId) {
-        if let VariantArmKind::BindAll { var } = kind {
-            // The whole-subject borrow: the subject's owner (an outer temp / param) keeps
-            // the reference; a consuming re-use in the arm (`err(e)`) Dups it.
-            self.value_of.insert(*var, subj);
-            self.param_values.insert(subj);
-            return;
-        }
-        if let VariantArmKind::Ctor { binds, .. } = kind {
-            for (slot, var, is_heap, fty) in binds {
-                let off = layout::slot_offset(*slot) as i64;
-                if *is_heap {
-                    // BORROW the slot handle: the subject owns the String; a move-out auto-Dups
-                    // in `lower_heap_result_arm`, a consuming re-use Dups in `lower_owned_heap_field`.
-                    let p = self.load_at_offset(h, off, crate::PrimKind::LoadHandle);
-                    self.param_values.insert(p);
-                    // An Option/Result PAYLOAD field (`Box(Option[Int])` — the depth-2
-                    // single-outer + builtin-inner class): seed the borrowed handle's
-                    // READ-shape so the inner `match $f { Some(n)/None }` executes (reads the
-                    // len/cap tag) instead of walling on a scalar destructure. Gated to
-                    // Option/Result exactly (the canonical seeder's variant arms) — a
-                    // String/List/nested-variant bind keeps today's borrow-only discipline.
-                    if matches!(fty,
-                        Ty::Applied(
-                            almide_lang::types::constructor::TypeConstructorId::Option
-                                | almide_lang::types::constructor::TypeConstructorId::Result,
-                            _
-                        ))
-                    {
-                        self.seed_variant_param(p, fty);
-                    }
-                    // A CLOSURE payload (`Run(f) => f()` — the variant-stored closure
-                    // class): the borrowed handle IS a closure block — admit it to the
-                    // dispatch set so the arm's `f(…)` lowers to `CallIndirect`.
-                    if matches!(fty, Ty::Fn { .. }) {
-                        self.closure_values.insert(p);
-                    }
-                    self.value_of.insert(*var, p);
-                } else {
-                    // Rung-5 variants slab: a SCALAR payload is a plain slot of the
-                    // subject block — target-neutral ListGetScalar (native `subj[slot]`).
-                    let _ = off;
-                    let idx = self.fresh_value();
-                    self.ops.push(Op::ConstInt { dst: idx, value: *slot as i64 });
-                    let payload = self.fresh_value();
-                    self.ops.push(Op::ListGetScalar { dst: payload, list: subj, idx });
-                    self.value_of.insert(*var, payload);
-                }
-            }
-        }
-    }
-
-    /// Lower a UNIT-result custom-variant `match` in STATEMENT position (ADT brick 3, the unit
-    /// sibling of [`Self::try_lower_custom_variant_match`]) — read the tag@slot0 and run only the
-    /// taken arm's EFFECTS. The subject is ALREADY materialized/borrowed by the caller (the
-    /// statement-`Match` entry), passed as `subject_value`.
-    ///
-    /// A custom variant must NEVER fall to the both-arms LINEARIZATION (that runs every arm's
-    /// effects = a silent miscompile — e.g. all three `println`s instead of one), so this returns
-    /// `Err` (WALL) on an out-of-subset arm rather than declining to the linearizer. Each arm
-    /// runs in a per-arm frame (`lower_branch_arm`), wrapped in `IfThen`/`Else`/`EndIf` markers
-    /// (no-ops in `verify_ownership`); the last arm / any wildcard is the unconditional else.
-    pub(crate) fn lower_custom_variant_unit_match(
-        &mut self,
-        subject_ty: &Ty,
-        subject_value: Option<ValueId>,
-        arms: &[IrMatchArm],
-    ) -> Result<(), LowerError> {
-        use crate::PrimKind;
-        let wall = |what: &str| {
-            Err(LowerError::Unsupported(format!(
-                "custom-variant statement match {what} cannot be faithfully lowered (a both-arms \
-                 linearization would run every arm's effects) not in this brick"
-            )))
-        };
-        let Some(subj) = subject_value else {
-            return wall("over a non-materialized subject");
-        };
-        // A DEFERRED-Opaque subject is an EMPTY block: reading its tag would execute a
-        // wrong arm silently (the record-ctor mt2 miscompile) — wall it honestly.
-        if self.deferred_opaque_binds.contains(&subj) {
-            return wall("over a deferred (unmaterialized) subject");
-        }
-        let type_name = match self.custom_variant_type_name(subject_ty) {
-            Some(n) => n,
-            None => return wall("over an unresolved variant type"),
-        };
-        let layout = match self.variant_layouts.by_type.get(&type_name) {
-            Some(l) => l.clone(),
-            None => return wall("over an unregistered variant"),
-        };
-        let plans = match self.parse_variant_arms(&layout, arms) {
-            Some(p) if !p.is_empty() => p,
-            _ => return wall("with an arm outside the scalar-field subset"),
-        };
-        let h = self.fresh_value();
-        self.ops.push(Op::Prim { kind: PrimKind::Handle, dst: Some(h), args: vec![subj] });
-        // Rung-5 variants slab: slot-0 tag via the target-neutral ListGetScalar.
-        let idx = self.fresh_value();
-        self.ops.push(Op::ConstInt { dst: idx, value: 0 });
-        let tag = self.fresh_value();
-        self.ops.push(Op::ListGetScalar { dst: tag, list: subj, idx });
-        self.emit_variant_unit_chain(h, tag, &plans, subj)
-    }
-
-    /// Emit the right-nested `if tag == t0 { arm0 } else if … else { last }` chain for a
-    /// UNIT-result custom-variant statement match. Each arm is a per-arm effect frame
-    /// (`lower_branch_arm` with no result), the markers carry `val: None`. The last plan / any
-    /// wildcard is the unconditional else. `Err` (the whole match walls) if an arm body is out
-    /// of subset — the unit sibling of [`Self::emit_variant_arm_chain`].
-    fn emit_variant_unit_chain(
-        &mut self,
-        h: ValueId,
-        tag: ValueId,
-        plans: &[(VariantArmKind, &IrExpr)],
-        subj: ValueId,
-    ) -> Result<(), LowerError> {
-        let Some(((kind, body), rest)) = plans.split_first() else {
-            return Ok(());
-        };
-        if rest.is_empty() || matches!(kind, VariantArmKind::Wildcard | VariantArmKind::BindAll { .. }) {
-            return self.lower_variant_unit_arm(kind, body, h, subj);
-        }
-        let arm_tag = match kind {
-            VariantArmKind::Ctor { tag, .. } => *tag,
-            VariantArmKind::Wildcard | VariantArmKind::BindAll { .. } => {
-                unreachable!("handled above")
-            }
-        };
-        let tc = self.fresh_value();
-        self.ops.push(Op::ConstInt { dst: tc, value: arm_tag });
-        let cond = self.fresh_value();
-        self.ops.push(Op::IntBinOp { dst: cond, op: IntOp::Eq, a: tag, b: tc });
-        self.ops.push(Op::IfThen { cond, dst: None });
-        self.lower_variant_unit_arm(kind, body, h, subj)?;
-        self.ops.push(Op::Else { val: None });
-        self.emit_variant_unit_chain(h, tag, rest, subj)?;
-        self.ops.push(Op::EndIf { val: None });
-        Ok(())
-    }
-
-    /// Lower one UNIT-statement custom-variant arm (its effects), with a PER-ARM FRAME that
-    /// drops the arm's `Dup`'d heap-field binds at arm end (the unit sibling of
-    /// [`Self::lower_variant_arm_value`]). The mark precedes `bind_variant_arm` so a heap field
-    /// bound + read by the effect (`println(s)`) is released here. Scalar arms add nothing.
-    fn lower_variant_unit_arm(
-        &mut self,
-        kind: &VariantArmKind,
-        body: &IrExpr,
-        h: ValueId,
-        subj: ValueId,
-    ) -> Result<(), LowerError> {
-        let mark = self.live_heap_handles.len();
-        self.bind_variant_arm(kind, h, subj);
-        // This arm executes under REAL `IfThen`/`Else`/`EndIf` markers (only the taken
-        // arm runs — `emit_variant_unit_chain`), so it is an EXECUTABLE unit arm:
-        // raise `unit_arm_depth` so a shared-cell/mutable-global write inside it is
-        // admitted as a real conditional effect (the modeled-frame guard keys on this;
-        // `lower_branch_arm` alone raises only `in_frame`, which also covers the
-        // both-arms linearization where such a write MUST wall).
-        self.unit_arm_depth += 1;
-        let r = self.lower_branch_arm(None, body);
-        self.unit_arm_depth -= 1;
-        r?;
-        self.drop_arm_locals(mark);
-        Ok(())
-    }
-
-    /// Emit the right-nested `if tag == t0 { arm0 } else if … else { last }` chain for a
-    /// custom-variant value match, returning the ValueId holding the chain's result. The LAST
-    /// plan is the unconditional `else` (no tag test — exhaustiveness guarantees it matches); a
-    /// `Wildcard` anywhere is likewise an unconditional `else` (the rest is unreachable). Each
-    /// arm body lowers in its own per-arm frame — `lower_scalar_arm` for a scalar result
-    /// (ADT brick 3), `lower_heap_result_arm` for a heap result (ADT brick 4, the arm moves out
-    /// a fresh heap value). `None` (caller rolls back) if an arm body is outside the subset.
-    fn emit_variant_arm_chain(
-        &mut self,
-        h: ValueId,
-        tag: ValueId,
-        plans: &[(VariantArmKind, &IrExpr)],
-        result_ty: &Ty,
-        subj: ValueId,
-    ) -> Option<ValueId> {
-        let heap = is_heap_ty(result_ty);
-        let ((kind, body), rest) = plans.split_first()?;
-        // The last arm, or any Wildcard, is the unconditional else (no tag test).
-        if rest.is_empty() || matches!(kind, VariantArmKind::Wildcard | VariantArmKind::BindAll { .. }) {
-            return self.lower_variant_arm_value(kind, body, h, result_ty, heap, subj);
-        }
-        let arm_tag = match kind {
-            VariantArmKind::Ctor { tag, .. } => *tag,
-            VariantArmKind::Wildcard | VariantArmKind::BindAll { .. } => {
-                unreachable!("handled above")
-            }
-        };
-        let dst = self.fresh_value();
-        let tc = self.fresh_value();
-        self.ops.push(Op::ConstInt { dst: tc, value: arm_tag });
-        let cond = self.fresh_value();
-        self.ops.push(Op::IntBinOp { dst: cond, op: IntOp::Eq, a: tag, b: tc });
-        self.ops.push(Op::IfThen { cond, dst: Some(dst) });
-        // RELEASE PARITY (mirrors lower_heap_result_if_inner): an OUTER handle
-        // this arm moves out must be released by the rest of the chain, and vice
-        // versa — otherwise the accounting is path-dependent (the branch-grouped
-        // cert `{m|}` rejects it; this keeps the lowering ahead of the checker).
-        let outer: Vec<ValueId> = self.live_heap_handles.clone();
-        let then_v = self.lower_variant_arm_value(kind, body, h, result_ty, heap, subj)?;
-        let consumed_by_then: Vec<ValueId> =
-            outer.iter().copied().filter(|x| !self.live_heap_handles.contains(x)).collect();
-        let else_marker_at = self.ops.len();
-        self.ops.push(Op::Else { val: Some(then_v) });
-        let live_after_then: Vec<ValueId> = self.live_heap_handles.clone();
-        let else_v = self.emit_variant_arm_chain(h, tag, rest, result_ty, subj)?;
-        let consumed_by_else: Vec<ValueId> = live_after_then
-            .iter()
-            .copied()
-            .filter(|x| !self.live_heap_handles.contains(x))
-            .collect();
-        for x in &consumed_by_then {
-            if !consumed_by_else.contains(x) {
-                let op = self.drop_op_for(*x);
-                self.ops.push(op); // the rest of the chain releases what this arm moved out
-            }
-        }
-        for x in &consumed_by_else {
-            if !consumed_by_then.contains(x) {
-                let op = self.drop_op_for(*x);
-                self.ops.insert(else_marker_at, op); // this arm releases what the chain moved out
-            }
-        }
-        self.ops.push(Op::EndIf { val: Some(else_v) });
-        Some(dst)
-    }
-
-    /// Lower one custom-variant arm to its value, with a PER-ARM FRAME that drops the arm's
-    /// `Dup`'d heap-field binds at arm end. The mark is taken BEFORE `bind_variant_arm` (whose
-    /// owned heap binds land in `live_heap_handles`), so `drop_arm_locals` releases exactly the
-    /// fields not moved out: a borrow-passed field (`tos(l)`) drops here; a moved-out field
-    /// (`Text(s) => s`) was `Dup`+`Consume`'d again by `lower_heap_result_arm`, so its original
-    /// bind still drops here (rc-balanced — the transient extra ref is freed). A scalar arm adds
-    /// nothing to the frame, so this is a no-op for the brick-2/3 paths.
-    fn lower_variant_arm_value(
-        &mut self,
-        kind: &VariantArmKind,
-        body: &IrExpr,
-        h: ValueId,
-        result_ty: &Ty,
-        heap: bool,
-        subj: ValueId,
-    ) -> Option<ValueId> {
-        let mark = self.live_heap_handles.len();
-        self.bind_variant_arm(kind, h, subj);
-        let v = if heap {
-            self.lower_heap_result_arm(body, result_ty)
-        } else {
-            self.lower_scalar_arm(body)
-        }?;
-        self.drop_arm_locals(mark);
-        Some(v)
-    }
 }
