+++ crates/almide-mir/src/lower/control_p3.rs
@@ -611,1139 +611,4 @@ impl LowerCtx {
         self.drop_arm_locals(lhh_mark);
         val
     }
-
-    /// Drop every heap handle the current scope frame added beyond `mark` (LIFO),
-    /// restoring `live_heap_handles` to its pre-frame length — the per-arm teardown.
-    pub(crate) fn drop_arm_locals(&mut self, mark: usize) {
-        // Release the handles created SINCE `mark`. Clamp to the current length: a sub-lowering
-        // (e.g. a match that drops its own subject) can leave `live_heap_handles` SHORTER than
-        // `mark`, in which case nothing new is live and this is a no-op (a bare `split_off(mark)`
-        // would panic). Semantically "drop everything from mark onward, nothing if none".
-        let mark = mark.min(self.live_heap_handles.len());
-        for v in self.live_heap_handles.split_off(mark).into_iter().rev() {
-            let op = self.drop_op_for(v);
-            self.ops.push(op);
-        }
-    }
-
-    /// Lower a `for v in iterable { body }` by modeling ONE iteration with a
-    /// PER-ITERATION SCOPE FRAME. Each iteration is internally balanced (its loop
-    /// variable + body locals are all dropped at iteration end), so N runtime
-    /// iterations are N balanced episodes — no cross-iteration leak or double-free,
-    /// and the flat cert (one iteration) is sound for any N (including 0: every op is
-    /// in a balanced frame). NO loop op — the iteration discipline lives entirely in
-    /// the lowering, the checker stays a flat fold.
-    ///
-    /// The ITERABLE is evaluated once: a heap iterable is lowered by `lower_call_args`
-    /// — an already-tracked `Var` is BORROWED, a FRESH heap value (a call/literal
-    /// result) is MATERIALIZED into an owned temp released at the OUTER scope; a scalar
-    /// iterable (a `Range`) carries no ownership. The LOOP VARIABLE binds one element per
-    /// iteration: a HEAP element ALIASES the whole container (`Op::Dup`, container-
-    /// grain like field extraction — it keeps the container alive for the iteration,
-    /// dropped at its end; element-precise identity needs the layout brick), a SCALAR
-    /// element is a `Const`. A `break`/`continue` is a no-op admitted ONLY over a
-    /// SCALAR-only frame (`wall_break_over_heap_frame`); over a heap frame it is WALLED
-    /// (a real early exit would skip a per-iteration heap Drop = a wasm leak). A HEAP
-    /// reassignment (the accumulator, `acc = acc + [x]`) is DEFERRED, not walled: the
-    /// `in_frame` discipline keeps `acc` pinned to its still-live handle across
-    /// iterations (memory-safe; the accumulation is deferred like every `Opaque`) and it
-    /// is not a frame handle. A scalar reassignment (`i = i + 1`) is a Copy `Const`,
-    /// harmless, admitted.
-    pub(crate) fn lower_for_in(
-        &mut self,
-        var: VarId,
-        var_tuple: &Option<Vec<VarId>>,
-        iterable: &IrExpr,
-        body: &[IrStmt],
-    ) -> Result<(), LowerError> {
-        // First try to EXECUTE a scalar `for i in start..end` as a real loop; out of that
-        // subset it rolls back and we keep the model-one-iteration form below.
-        if self.try_lower_scalar_for_range(var, var_tuple, iterable, body) {
-            return Ok(());
-        }
-        // Then try to EXECUTE `for x in xs` over a List[Int] as a real element loop.
-        if self.try_lower_scalar_for_list(var, var_tuple, iterable, body) {
-            return Ok(());
-        }
-        // Then `for (k, v) in m` / `for k in m` over a self-hosted Map layout as a
-        // real entry loop.
-        if self.try_lower_scalar_for_map(var, var_tuple, iterable, body) {
-            return Ok(());
-        }
-        // The iterable is evaluated ONCE before the loop. A heap iterable goes through
-        // `lower_call_args` — an already-tracked `Var` is borrowed (no new ownership),
-        // a fresh heap value is materialized into an owned temp dropped at the OUTER
-        // scope (its caps captured by the lowering). A scalar iterable (a `Range`)
-        // carries no ownership; capture any call in it for caps.
-        let container: Option<ValueId> = if is_heap_ty(&iterable.ty)
-            // A Range ITERABLE stays on the no-container model path (it carries no
-            // ownership; the call-arg Range materialization emits a `list.range` CallFn
-            // the caps ledger only accounts for in ARGUMENT position).
-            && !matches!(&iterable.kind, IrExprKind::Range { .. })
-        {
-            match self.lower_call_args(std::slice::from_ref(iterable))?.into_iter().next() {
-                Some(CallArg::Handle(v)) => Some(v),
-                _ => None,
-            }
-        } else {
-            self.record_elided_calls(iterable);
-            None
-        };
-        let mark = self.live_heap_handles.len();
-        let vars: Vec<VarId> = match var_tuple {
-            Some(vs) => vs.clone(),
-            None => vec![var],
-        };
-        for v in vars {
-            // A heap element aliases the whole container; a scalar element is a Const.
-            let elem_ty = find_var_ty(body, v);
-            let elem_heap = elem_ty.as_ref().map(|t| is_heap_ty(t)).unwrap_or(false);
-            if elem_heap {
-                // The model-one-iteration fallback aliases the element to the WHOLE container (a `Dup`
-                // of the list handle) — the real per-element loop (`try_lower_scalar_for_list`) only
-                // covers List[scalar]/List[String], which never reach here. A heap-AGGREGATE element
-                // (tuple/record) DOES reach here, and reading a field of it (`for p in ps { p.0 }`)
-                // would read the CONTAINER's slot, not the element's — a SILENT MISCOMPILE. Now that
-                // the producer side materializes List[tuple]/List[record] returns, that consumer is
-                // reachable, so WALL it (honest) until the heap-aggregate per-element borrow lands.
-                if elem_ty.as_ref().is_some_and(|t| self.aggregate_field_tys(t).is_some())
-                    && body_reads_var_field(body, v)
-                {
-                    return Err(LowerError::Unsupported(
-                        "for-in over a List of heap aggregates (tuple/record) whose element is read \
-                         via a direct field/index (`for p in ps { p.0 }`) is not in this brick: the \
-                         model-one-iteration fallback aliases the element to the whole container, so \
-                         the projection would read the container — walled to avoid a silent \
-                         miscompile (needs the heap-aggregate per-element borrow)"
-                            .into(),
-                    ));
-                }
-                let src = container.ok_or_else(|| {
-                    LowerError::Unsupported(
-                        "for-in heap loop variable over a non-container iterable not in this brick".into(),
-                    )
-                })?;
-                let dst = self.fresh_value();
-                self.ops.push(Op::Dup { dst, src });
-                self.value_of.insert(v, dst);
-                self.live_heap_handles.push(dst);
-            } else {
-                if crate::lower::strict_values() {
-                    return Err(crate::lower::strict_const_wall("for-in loop variable"));
-                }
-                let dst = self.fresh_value();
-                self.ops.push(Op::Const { dst });
-                self.value_of.insert(v, dst);
-            }
-        }
-        // A heap reassignment in the body is the loop ACCUMULATOR (`acc = acc + [x]`):
-        // it is DEFERRED, not rebound (the `in_frame` discipline) — `acc` keeps its
-        // still-live handle across iterations, so the next iteration never dereferences
-        // a freed handle. Memory-safe; the accumulation itself is deferred like `Opaque`.
-        self.in_frame += 1;
-        for stmt in body {
-            self.lower_stmt(stmt)?;
-        }
-        self.in_frame -= 1;
-        self.wall_break_over_heap_frame(body, "for-in", mark)?;
-        self.drop_arm_locals(mark);
-        Ok(())
-    }
-
-    /// Lower a `while cond { body }` like a `for-in` body — a PER-ITERATION SCOPE
-    /// FRAME makes one modeled iteration balanced, sound for any N. The condition is
-    /// evaluated each iteration (its caps captured); the body's locals are dropped at
-    /// iteration end. Same as `for-in`: a `break`/`continue` over a HEAP frame is walled
-    /// (post-lowering), a heap reassignment (accumulator) deferred by `in_frame`.
-    /// Try to lower `while cond { body }` as a REAL scalar-state loop that EXECUTES N
-    /// times (the `LoopStart`/`LoopBreakUnless`/`LoopEnd` markers), reassigning scalar
-    /// loop-carried state via [`Op::SetLocal`]. Restricted to the sound + runnable subset:
-    /// an Int/Bool cond, NO `break`/`continue` (a no-op early-exit would be wrong inside a
-    /// real loop), and a body with NO heap reassignment (the `scalar_loop_depth` Assign
-    /// rule errors on one) and NO net heap handle escaping the per-iteration frame. The
-    /// cond ops sit INSIDE the loop (re-evaluated each iteration); per-iteration heap (a
-    /// string literal in `println`) is dropped before the back-edge. SOUNDNESS by REUSE:
-    /// the markers are no-ops in verify_ownership and the body is a per-iteration-balanced
-    /// frame — the cert verifies ONE balanced iteration, sound for any N (the existing
-    /// model-one-iteration argument), the markers only make wasm actually run it N times.
-    /// Returns false (and rolls back) when out of subset; `lower_while` then falls back.
-    /// Pre-loop OWNED COPY for every borrowed-param slot the loop body HEAP-REASSIGNS
-    /// (the C-132 callee shape after the move-mode rewrite: `fn addc(mut s: String, n)
-    /// = { while k < n { s = s + "x"; … } … }` — the rebind arrives via the string.push
-    /// functional rewrite). The loop rebind's uniform drop-old would free the CALLER's
-    /// buffer on iteration 1 (the mut_heap_param rc-underflow trap); starting the slot
-    /// as a `Dup` (+1) of the borrowed param makes iteration 1 drop the copy (the
-    /// caller's reference stays live) and later iterations drop the owned
-    /// intermediates — the same accounting the TCO pre-copy proves. The Dup joins
-    /// `live_heap_handles` (scope-end drops the FINAL object through the same local);
-    /// its cert `a` is backed by the real `Op::Dup` (the borrow-by-default gate).
-    pub(crate) fn precopy_borrowed_reassign_slots(&mut self, body: &[IrStmt]) {
-        let mut vars: Vec<VarId> = Vec::new();
-        collect_heap_reassign_vars(body, &mut vars);
-        for var in vars {
-            if let Some(&val) = self.value_of.get(&var) {
-                if self.param_values.contains(&val) {
-                    let owned = self.fresh_value();
-                    self.ops.push(Op::Dup { dst: owned, src: val });
-                    self.value_of.insert(var, owned);
-                    self.live_heap_handles.push(owned);
-                }
-            }
-        }
-    }
-
-    pub(crate) fn try_lower_scalar_while(&mut self, cond: &IrExpr, body: &[IrStmt]) -> bool {
-        if !matches!(cond.ty, Ty::Int | Ty::Bool) {
-            return false;
-        }
-        let ops_mark = self.ops.len();
-        let lhh_mark = self.live_heap_handles.len();
-        let lifted_mark = self.lifted.len();
-        let value_of_snapshot = self.value_of.clone();
-        self.precopy_borrowed_reassign_slots(body);
-
-        self.ops.push(Op::LoopStart);
-        let cond_v = match self.lower_scalar_value(cond) {
-            Some(v) => v,
-            None => {
-                self.rollback_scalar_loop(ops_mark, lhh_mark, lifted_mark, value_of_snapshot);
-                return false;
-            }
-        };
-        self.ops.push(Op::LoopBreakUnless { cond: cond_v });
-
-        let body_mark = self.live_heap_handles.len();
-        self.in_frame += 1;
-        self.scalar_loop_depth += 1;
-        let mut ok = true;
-        for stmt in body {
-            if self.lower_while_body_stmt(stmt).is_err() {
-                ok = false;
-                break;
-            }
-        }
-        self.scalar_loop_depth -= 1;
-        self.in_frame -= 1;
-
-        if !ok {
-            self.rollback_scalar_loop(ops_mark, lhh_mark, lifted_mark, value_of_snapshot);
-            return false;
-        }
-        // Per-iteration heap (a string literal in a body `println`) is released before the
-        // back-edge, INSIDE the loop, so each iteration is balanced.
-        self.drop_arm_locals(body_mark);
-        self.ops.push(Op::LoopEnd);
-        true
-    }
-
-    /// One while-body statement, admitting the CONDITIONAL-BREAK forms the real loop
-    /// can execute with the EXISTING marker vocabulary (no new op):
-    ///   `if c then <rest> else break`  (the guard-else-break desugar) →
-    ///       `LoopBreakUnless(c)` then <rest> emitted linearly — on the broken path the
-    ///       `br` already exited, exactly like the loop-head condition;
-    ///   `if c then break else ()`      (the do-block shape) →
-    ///       `LoopBreakUnless(1 - c)`.
-    /// Any OTHER Break/Continue in the statement ERRS (aborting the attempt → the
-    /// model-one-iteration fallback then WALLS): `lower_stmt` silently swallows a bare
-    /// Break (mod_p3), so delegating one would silently drop the early exit.
-    /// `{ A…; break }` → `Some({ A… })` — the then-arm of a mid-body conditional break
-    /// with statements BEFORE the break. `None` when the expr does not end in a bare
-    /// trailing break (or has nothing before it — the simpler cases own that shape).
-    fn strip_trailing_break_expr(e: &IrExpr) -> Option<IrExpr> {
-        let IrExprKind::Block { stmts, expr } = &e.kind else { return None };
-        // Trailing break as the block TAIL.
-        if let Some(t) = expr.as_deref() {
-            if matches!(t.kind, IrExprKind::Break) && !stmts.is_empty() {
-                return Some(IrExpr {
-                    kind: IrExprKind::Block { stmts: stmts.clone(), expr: None },
-                    ty: almide_lang::types::Ty::Unit,
-                    span: e.span.clone(),
-                    def_id: e.def_id,
-                });
-            }
-            return None;
-        }
-        // Trailing break as the LAST statement.
-        if stmts.len() >= 2 {
-            if let IrStmtKind::Expr { expr: last } = &stmts[stmts.len() - 1].kind {
-                if matches!(last.kind, IrExprKind::Break) {
-                    return Some(IrExpr {
-                        kind: IrExprKind::Block {
-                            stmts: stmts[..stmts.len() - 1].to_vec(),
-                            expr: None,
-                        },
-                        ty: almide_lang::types::Ty::Unit,
-                        span: e.span.clone(),
-                        def_id: e.def_id,
-                    });
-                }
-            }
-        }
-        None
-    }
-
-    fn lower_while_body_stmt(&mut self, stmt: &IrStmt) -> Result<(), LowerError> {
-        fn is_break(e: &IrExpr) -> bool {
-            match &e.kind {
-                IrExprKind::Break => true,
-                IrExprKind::Block { stmts, expr } => {
-                    (stmts.is_empty() && expr.as_deref().is_some_and(is_break))
-                        || (expr.is_none()
-                            && stmts.len() == 1
-                            && matches!(&stmts[0].kind,
-                                IrStmtKind::Expr { expr } if is_break(expr)))
-                }
-                _ => false,
-            }
-        }
-        fn is_unit(e: &IrExpr) -> bool {
-            match &e.kind {
-                IrExprKind::Unit => true,
-                IrExprKind::Block { stmts, expr } => {
-                    stmts.is_empty() && expr.as_deref().map_or(true, is_unit)
-                }
-                _ => false,
-            }
-        }
-        if let IrStmtKind::Expr { expr } = &stmt.kind {
-            // A BARE break statement (`if true then break else ()` const-folds to it):
-            // break unconditionally — LoopBreakUnless over a 0 cond.
-            if matches!(expr.kind, IrExprKind::Break) {
-                let z = self.fresh_value();
-                self.ops.push(Op::ConstInt { dst: z, value: 0 });
-                self.ops.push(Op::LoopBreakUnless { cond: z });
-                return Ok(());
-            }
-            if let IrExprKind::If { cond, then, else_ } = &expr.kind {
-                if is_break(else_) {
-                    let c = self.lower_scalar_value(cond).ok_or_else(|| {
-                        LowerError::Unsupported("while conditional-break cond".into())
-                    })?;
-                    self.ops.push(Op::LoopBreakUnless { cond: c });
-                    return self.lower_while_body_inline(then);
-                }
-                if is_break(then) && is_unit(else_) {
-                    let c = self.lower_scalar_value(cond).ok_or_else(|| {
-                        LowerError::Unsupported("while conditional-break cond".into())
-                    })?;
-                    let one = self.fresh_value();
-                    self.ops.push(Op::ConstInt { dst: one, value: 1 });
-                    let nc = self.fresh_value();
-                    self.ops.push(Op::IntBinOp { dst: nc, op: IntOp::Sub, a: one, b: c });
-                    self.ops.push(Op::LoopBreakUnless { cond: nc });
-                    return Ok(());
-                }
-                // `if c then { A…; break } else B` (find_factor: `if n % i == 0 then
-                // { result = i; break } else { i = i + 1 }`): CAPTURE the (call-free,
-                // pure-scalar) cond once, run the ordinary unit `if` with the trailing
-                // break STRIPPED (both arms then break-free — the statement-if machinery
-                // branches the arm assigns correctly), and break on the CAPTURED value
-                // after. The capture keeps the break test the value the branch dispatched
-                // on even when an arm mutates the cond's operands (`i = i + 1`).
-                if let Some(then_stripped) = Self::strip_trailing_break_expr(then) {
-                    if !body_breaks_or_continues(std::slice::from_ref(&IrStmt {
-                        kind: IrStmtKind::Expr { expr: then_stripped.clone() },
-                        span: stmt.span.clone(),
-                    })) && !body_breaks_or_continues(std::slice::from_ref(&IrStmt {
-                        kind: IrStmtKind::Expr { expr: (**else_).clone() },
-                        span: stmt.span.clone(),
-                    })) && !crate::lower::expr_contains_call(cond)
-                    {
-                        let c = self.lower_scalar_value(cond).ok_or_else(|| {
-                            LowerError::Unsupported("while conditional-break cond".into())
-                        })?;
-                        let unit_if = IrStmt {
-                            kind: IrStmtKind::Expr {
-                                expr: IrExpr {
-                                    kind: IrExprKind::If {
-                                        cond: cond.clone(),
-                                        then: Box::new(then_stripped),
-                                        else_: else_.clone(),
-                                    },
-                                    ty: almide_lang::types::Ty::Unit,
-                                    span: expr.span.clone(),
-                                    def_id: expr.def_id,
-                                },
-                            },
-                            span: stmt.span.clone(),
-                        };
-                        self.lower_stmt(&unit_if)?;
-                        let one = self.fresh_value();
-                        self.ops.push(Op::ConstInt { dst: one, value: 1 });
-                        let nc = self.fresh_value();
-                        self.ops.push(Op::IntBinOp { dst: nc, op: IntOp::Sub, a: one, b: c });
-                        self.ops.push(Op::LoopBreakUnless { cond: nc });
-                        return Ok(());
-                    }
-                }
-            }
-        }
-        if body_breaks_or_continues(std::slice::from_ref(stmt)) {
-            return Err(LowerError::Unsupported(
-                "unrecognized break/continue in a while body (lower_stmt would silently \
-                 swallow it) not in this brick"
-                    .into(),
-            ));
-        }
-        self.lower_stmt(stmt)
-    }
-
-    /// Inline the surviving arm of a while conditional-break (`if c then <rest> else
-    /// break`): a Block's statements re-enter [`Self::lower_while_body_stmt`] (a nested
-    /// conditional break chains), a unit tail is nothing, any other tail lowers as an
-    /// effect statement.
-    fn lower_while_body_inline(&mut self, e: &IrExpr) -> Result<(), LowerError> {
-        match &e.kind {
-            IrExprKind::Unit => Ok(()),
-            IrExprKind::Block { stmts, expr } => {
-                for s in stmts {
-                    self.lower_while_body_stmt(s)?;
-                }
-                match expr.as_deref() {
-                    Some(t) => self.lower_while_body_inline(t),
-                    None => Ok(()),
-                }
-            }
-            _ => self.lower_while_body_stmt(&IrStmt {
-                kind: IrStmtKind::Expr { expr: e.clone() },
-                span: e.span.clone(),
-            }),
-        }
-    }
-
-    /// Roll back a scalar-loop ATTEMPT (`try_lower_scalar_while` / `_for_range` / `_for_list`),
-    /// restoring EVERY side-effect the partial body lowering may have produced — not only `ops`
-    /// but the LAMBDA-LIFTED auxiliaries (`self.lifted`). A lambda call-arg in the body (`for x in
-    /// xs { … list.map([y], (u) => …) … }`) lifts a `__lambda_*` MirFunction into `self.lifted`;
-    /// if the attempt then rolls back (a heap reassignment aborts it → the model-one-iteration
-    /// fallback re-lowers the SAME body, re-lifting the lambda), the abandoned first copy would
-    /// survive and DOUBLE-COUNT its inner calls (a `mir > ir` wall breach). Truncating `lifted` to
-    /// `lifted_mark` (captured at THIS attempt's start, threaded as a local so NESTED loop attempts
-    /// each roll back to their own floor) makes the rollback total.
-    fn rollback_scalar_loop(
-        &mut self,
-        ops_mark: usize,
-        lhh_mark: usize,
-        lifted_mark: usize,
-        value_of_snapshot: std::collections::HashMap<almide_ir::VarId, ValueId>,
-    ) {
-        self.ops.truncate(ops_mark);
-        self.live_heap_handles.truncate(lhh_mark);
-        self.lifted.truncate(lifted_mark);
-        self.value_of = value_of_snapshot;
-    }
-
-    /// Try to lower a HEAP-RESULT `if cond then A else B` (a String/data-returning branch)
-    /// to EXECUTABLE control flow — only the taken arm allocates, and its value is the
-    /// function result. SOUNDNESS by PER-ARM BALANCE (no Coq change — see
-    /// docs/roadmap/active/v1-heap-result-control-flow.md): each arm `Alloc`s its value
-    /// (cert `i`) AND `Consume`s it (cert `m`) so the arm is internally `"im"` balanced
-    /// exactly like a scalar arm carries none; the `IfThen` result `dst` is NEVER an
-    /// `Alloc`, so it is not in the ownership cert's object set and `func.ret = dst` emits
-    /// no second move-out (no double-free). The render selects one arm at runtime
-    /// (`(if (result i32) …)`), so exactly one `Alloc` happens and is returned rc=1 to the
-    /// caller — the untaken arm never allocates (no leak). FIRST version: both arms are
-    /// direct string LITERALS (the common `if c then "a" else "b"`); other arm kinds fall
-    /// back to today's sound Opaque form. Returns the result `dst`, or `None` (rolled
-    /// back) when out of subset. Arms may be string LITERALS or a NESTED heap-result `if`
-    /// (the else-if chain a desugared `match` produces — `match n { 0 => "a", _ => "b" }`),
-    /// recursively. Other arm kinds fall back to today's sound Opaque form.
-    pub(crate) fn try_lower_heap_result_if(
-        &mut self,
-        cond: &IrExpr,
-        then: &IrExpr,
-        else_: &IrExpr,
-        result_ty: &Ty,
-    ) -> Option<ValueId> {
-        if !is_heap_ty(result_ty) {
-            return None;
-        }
-        // The whole attempt rolls back as a unit: the recursion below truncates nothing,
-        // so the OUTERMOST call restores the op stream AND the live-handle set on any
-        // out-of-subset arm (a call arm may have materialized + tracked a temp).
-        let ops_mark = self.ops.len();
-        let lhh_mark = self.live_heap_handles.len();
-        let result = self.lower_heap_result_if_inner(cond, then, else_, result_ty);
-        if result.is_none() {
-            self.ops.truncate(ops_mark);
-            self.live_heap_handles.truncate(lhh_mark);
-        }
-        result
-    }
-
-    /// Materialize the CONDITION of a heap-result `if` to a scalar (Bool = i64 0/1)
-    /// BEFORE the `IfThen` marker. The common shape is a pure `lower_scalar_value` cond
-    /// (a comparison, a Var, a literal) — tried first, no ownership. When that defers, a
-    /// Bool/Int-returning PURE call WITH HEAP ARGS (`if string.contains(s, x) then …`,
-    /// `if list.is_empty(xs) then …`) is materialized via `try_lower_scalar_call`: the
-    /// call's heap arg temps are pushed to `live_heap_handles`, and a per-cond frame
-    /// (`drop_arm_locals`) frees them IMMEDIATELY after the call — they are transient to
-    /// the condition, not owned by either arm. The scalar result is not a heap handle, so
-    /// it survives the frame teardown. SOUND: the cond eval is internally balanced (each
-    /// arg temp alloc'd `i` + dropped `d` within the frame), exactly the per-arm
-    /// discipline; outside the pure-scalar-call subset it walls (`None` → Opaque). The
-    /// gate keeping a heap-arg call OUT of `lower_scalar_value` (its rollback-safe, no-
-    /// ownership contract) does not bind here — this position freely emits ownership ops.
-    fn lower_heap_result_cond(&mut self, cond: &IrExpr) -> Option<ValueId> {
-        // A scalar cond can itself MATERIALIZE a transient heap temp — `if c == "#"` lowers to
-        // `string.eq(c, "#")` whose `"#"` literal is a fresh owned String. That temp is dead the
-        // instant the Bool is computed, so it MUST be freed HERE, within a cond-local frame, BEFORE
-        // the `IfThen` marker — never deferred to the enclosing arm's `drop_arm_locals`. The cond
-        // of a NESTED heap-result `if` (the else-of-an-else parser shape) sits inside one arm's
-        // wasm branch; deferring its temp's `Drop` to the OUTER block scope emits an UNCONDITIONAL
-        // `rc_dec` of a local that the sibling arm never initialized (garbage/0 → trap). The frame
-        // keeps the cond internally `i…d`-balanced exactly where it executes.
-        let frame = self.live_heap_handles.len();
-        if let Some(v) = self.lower_scalar_value(cond) {
-            self.drop_arm_locals(frame);
-            return Some(v);
-        }
-        // A scalar-returning (Bool/Int) PURE call with heap args — materialize it, then
-        // free the transient arg temps within a cond-local frame.
-        if let IrExprKind::Call { .. } = &cond.kind {
-            if !is_heap_ty(&cond.ty) {
-                if let Some(v) = self.try_lower_scalar_call(cond, &cond.ty) {
-                    self.drop_arm_locals(frame);
-                    return Some(v);
-                }
-            }
-        }
-        // A Bool-returning `==`/`!=` over HEAP operands neither of which is a tracked Var
-        // (`list.first(xs) == some("")`, `string.first(s) == some(c)` — the toml `emit_sections`
-        // blocker). `lower_scalar_value` above already covers the case where BOTH operands are
-        // materialized Vars (its Option/Result eq reads them via `materialized_option_handle`);
-        // here the operands are a heap-returning CALL and/or an Option/Result CTOR, which it
-        // declines. Materialize EACH operand into an owned block in the SAME cond-local frame, run
-        // the typed eq over their handles, then `drop_arm_locals(frame)` frees the operand temps —
-        // they are transient to the condition, owned by NEITHER arm (the per-cond `i…d` balance,
-        // exactly the pure-call-cond path's discipline). A non-materializable operand returns None
-        // → the `if` keeps its sound Opaque wall (no invalid wasm, no leak).
-        if let IrExprKind::BinOp { op: op @ (almide_ir::BinOp::Eq | almide_ir::BinOp::Neq), left, right } = &cond.kind {
-            if matches!(cond.ty, Ty::Bool) && is_heap_ty(&left.ty) {
-                if let Some(v) = self.lower_heap_eq_cond(left, right, matches!(op, almide_ir::BinOp::Neq)) {
-                    self.drop_arm_locals(frame);
-                    return Some(v);
-                }
-                // On decline, the partial ops/handles a half-materialized operand left are restored
-                // by the OUTER `try_lower_heap_result_if` rollback (ops + live_heap_handles truncated
-                // to its marks on any `None`), so the caller's Opaque fallback starts clean — no
-                // extra teardown needed here.
-            }
-        }
-        None
-    }
-
-    /// Lower a HEAP `left == right` / `left != right` cond (Bool) whose operands are NOT both
-    /// tracked Vars — by MATERIALIZING each operand into an owned block (tracked in the cond
-    /// frame so `drop_arm_locals` frees it after the compare) and running the typed eq over the
-    /// two block handles. Reuses the SAME eq cores the Var-operand path uses
-    /// (`option_scalar_eq_from_handles` / `option_heap_eq_from_handles` /
-    /// `result_scalar_eq_from_handles`, `string.eq`, `value.eq`, `list.eq_*`) — no new eq op.
-    /// `negate` wraps the result as `1 - eq` (for `!=`). SELF-CONTAINED ROLLBACK: on any decline a
-    /// half-materialized operand may have emitted ops + pushed an owned temp to `live_heap_handles`;
-    /// this restores BOTH to the pre-attempt marks so the method leaves NO trace on `None` (the
-    /// rollback-safe contract the cond position requires — `lower_heap_result_cond` is reached from
-    /// paths that do NOT all wrap it in an outer truncate, e.g. the defunc HOF body). Returns None
-    /// (the caller walls) for an operand shape we cannot materialize, or a type whose eq is unhandled.
-    fn lower_heap_eq_cond(&mut self, left: &IrExpr, right: &IrExpr, negate: bool) -> Option<ValueId> {
-        let ops_mark = self.ops.len();
-        let lhh_mark = self.live_heap_handles.len();
-        let ty = &left.ty;
-        let eq = match self.lower_heap_eq_typed_materialized(left, right, ty) {
-            Some(eq) => eq,
-            None => {
-                self.ops.truncate(ops_mark);
-                self.live_heap_handles.truncate(lhh_mark);
-                return None;
-            }
-        };
-        if !negate {
-            return Some(eq);
-        }
-        let one = self.fresh_value();
-        self.ops.push(Op::ConstInt { dst: one, value: 1 });
-        let dst = self.fresh_value();
-        self.ops.push(Op::IntBinOp { dst, op: IntOp::Sub, a: one, b: eq });
-        Some(dst)
-    }
-
-    /// Materialize both operands of a heap `==` into owned blocks (in the current cond frame) and
-    /// emit the typed equality, returning the Bool ValueId. Each operand is materialized by
-    /// `materialize_eq_operand` (a tracked Var is BORROWED, a fresh heap value is an owned temp
-    /// added to `live_heap_handles` with its recursive drop set). The eq BORROWS the operand
-    /// handles (it only reads), so the owned temps survive to the frame teardown. The per-type
-    /// dispatch is the ONE recursive engine [`Self::typed_slot_eq`]; an unsupported shape returns
-    /// None and the caller rolls back (walls) — never wrong bytes.
-    pub(crate) fn lower_heap_eq_typed_materialized(
-        &mut self,
-        left: &IrExpr,
-        right: &IrExpr,
-        ty: &Ty,
-    ) -> Option<ValueId> {
-        let lb = self.materialize_eq_operand(left, ty)?;
-        let rb = self.materialize_eq_operand(right, ty)?;
-        let lh = self.handle_of(lb);
-        let rh = self.handle_of(rb);
-        self.typed_slot_eq(lh, rh, ty, 0)
-    }
-
-    /// Recursive typed equality over two 8-byte SLOT VALUES of type `ty` — the ONE eq engine
-    /// every `==`/`!=` compose routes through (the unit-if cond, the value-position BinOp, and
-    /// every nested payload/field recursion). A scalar slot holds the value itself (int-class
-    /// i64 / float-class f64 bits); a heap slot holds the block's byte-address HANDLE, read
-    /// through its layout:
-    ///   String/Value/List[T] → the borrowed module eq call; Option/Result → the tag-compare
-    ///   cores; tuple/record → per-slot recursion ANDed; custom variant → tag eq + a
-    ///   tag-dispatched per-field recursion chain.
-    /// Every path only READS (borrows) — no ownership events — so the recursion composes freely
-    /// inside branch merges, and a `None` anywhere propagates to the caller's rollback (the
-    /// unlowered shape walls, never wrong output). `depth` caps type-level recursion (a variant
-    /// containing itself — `Cons(Int, MyList)` — would otherwise recurse forever at COMPILE
-    /// time); a capped shape walls, honest (runtime-recursive eq needs a synthesized fn brick).
-    /// The `String`/`Value`/`List[T]` borrowed-handle module-eq call name for
-    /// [`Self::typed_slot_eq`]'s slot classification — a pure, `self`-free lookup keyed on
-    /// `ty` alone (no ownership events, matching the doc above). Verbatim extraction
-    /// (guard-clause flattening) of the former inline if-else-if chain, no behavior change —
-    /// see docs/roadmap/active/code-health-codopsy.md.
-    fn module_eq_call_name(ty: &Ty) -> Option<&'static str> {
-        use almide_lang::types::constructor::TypeConstructorId as TC;
-        if matches!(ty, Ty::String) {
-            return Some("string.eq");
-        }
-        if crate::lower::is_value_ty(ty) {
-            return Some("value.eq");
-        }
-        let Ty::Applied(TC::List, es) = ty else {
-            return None;
-        };
-        if es.len() != 1 {
-            return None;
-        }
-        Self::list_elem_eq_call_name(&es[0])
-    }
-
-    /// The `List[T]` ELEMENT classification half of [`Self::module_eq_call_name`] — split out
-    /// (rather than left nested) so neither function's guard-clause chain re-exceeds the
-    /// depth this extraction exists to fix. Verbatim extraction, no behavior change.
-    fn list_elem_eq_call_name(elem_ty: &Ty) -> Option<&'static str> {
-        use almide_lang::types::constructor::TypeConstructorId as TC;
-        if matches!(elem_ty, Ty::Int) {
-            return Some("list.eq_int");
-        }
-        if matches!(elem_ty, Ty::String) {
-            return Some("list.eq_str");
-        }
-        if matches!(elem_ty, Ty::Float) {
-            return Some("list.eq_float");
-        }
-        if matches!(elem_ty, Ty::Bool) {
-            return Some("list.eq_bool");
-        }
-        if crate::lower::is_value_ty(elem_ty) {
-            return Some("list.eq_value");
-        }
-        if let Ty::Applied(TC::List, inner) = elem_ty {
-            // Nested lists — the element-wise recursion into the flat list eq
-            // (value_core `list_eq_list_*`).
-            if inner.len() == 1 && matches!(inner[0], Ty::Int) {
-                return Some("list.eq_list_int");
-            }
-            if inner.len() == 1 && matches!(inner[0], Ty::Float) {
-                return Some("list.eq_list_float");
-            }
-            if inner.len() == 1 && matches!(inner[0], Ty::String) {
-                return Some("list.eq_list_str");
-            }
-            return None;
-        }
-        if let Ty::Applied(TC::Option, inner) = elem_ty {
-            // List[Option[Int/Bool]] — element-wise len-as-tag + i64 payload
-            // compare (value_core `list_eq_opt_int`). Scalar payloads only: a
-            // Float payload's slot is f64 BITS (bit-eq ≠ `==` on -0.0/NaN).
-            if inner.len() == 1 && matches!(inner[0], Ty::Int | Ty::Bool) {
-                return Some("list.eq_opt_int");
-            }
-            return None;
-        }
-        if let Ty::Tuple(ts) = elem_ty {
-            // List[Tuple[scalar…]] — a scalar tuple block is LAYOUT-IDENTICAL to a
-            // same-arity List of its slot class (len@4 = arity, 8-byte slots @12),
-            // so the nested-list eq of the matching class compares it exactly:
-            // Int/Bool slots bit-compare (list.eq_list_int), all-Float slots
-            // float-compare per slot (list.eq_list_float). A MIXED Int/Float
-            // tuple has no matching flat class (a bit-compare on the Float slot
-            // is wrong on -0.0/NaN) — decline, the eq site walls honestly.
-            if !ts.is_empty() && ts.iter().all(|t| matches!(t, Ty::Int | Ty::Bool)) {
-                return Some("list.eq_list_int");
-            }
-            if !ts.is_empty() && ts.iter().all(|t| matches!(t, Ty::Float)) {
-                return Some("list.eq_list_float");
-            }
-            return None;
-        }
-        None
-    }
-
-    pub(crate) fn typed_slot_eq(
-        &mut self,
-        lv: ValueId,
-        rv: ValueId,
-        ty: &Ty,
-        depth: u32,
-    ) -> Option<ValueId> {
-        use almide_lang::types::constructor::TypeConstructorId as TC;
-        const MAX_EQ_DEPTH: u32 = 8;
-        if depth > MAX_EQ_DEPTH {
-            return None;
-        }
-        // Scalars — the slot IS the value.
-        if Self::float_operand_ty(ty) {
-            let dst = self.fresh_value();
-            self.ops.push(Op::Prim {
-                kind: crate::PrimKind::FloatCmp(crate::FCmpOp::Eq),
-                dst: Some(dst),
-                args: vec![lv, rv],
-            });
-            return Some(dst);
-        }
-        if Self::int_eq_operand_ty(ty) {
-            let dst = self.fresh_value();
-            self.ops.push(Op::IntBinOp { dst, op: IntOp::Eq, a: lv, b: rv });
-            return Some(dst);
-        }
-        // String / Value / List[T] — the borrowed-handle module eq call.
-        let module_eq: Option<&str> = Self::module_eq_call_name(ty);
-        if let Some(name) = module_eq {
-            let dst = self.fresh_value();
-            self.ops.push(Op::CallFn {
-                dst: Some(dst),
-                name: name.to_string(),
-                args: vec![CallArg::Handle(lv), CallArg::Handle(rv)],
-                result: Some(repr_of(&Ty::Bool).ok()?),
-            });
-            return Some(dst);
-        }
-        // A `List[<custom variant>]` — the synthesized loop helper (the element
-        // eq is the variant helper; a non-variant element stayed in the module
-        // table above). Generated once per parent fn, called at the site.
-        if let Ty::Applied(TC::List, es) = ty {
-            if es.len() == 1 {
-                if let Some(elem_name) = self.custom_variant_type_name(&es[0]) {
-                    if let Some(elem_layout) =
-                        self.variant_layouts.by_type.get(&elem_name).cloned()
-                    {
-                        let (elem_key, elem_inst) =
-                            instantiate_variant_layout(&elem_name, &elem_layout, &es[0]);
-                        if self.ensure_list_eq_helper(&elem_key, &elem_inst) {
-                            let name = self.list_eq_helper_name(&elem_key);
-                            return Some(self.emit_eq_helper_call(name, lv, rv));
-                        }
-                    }
-                    return None;
-                }
-            }
-        }
-        // A custom VARIANT — a RECURSIVE one (or one already being generated)
-        // routes through its synthesized helper (recursion-by-call; the static
-        // inline could never terminate); everything else keeps the proven
-        // inline tag-dispatch chain, byte-identical to before.
-        if let Some(tyname) = self.custom_variant_type_name(ty) {
-            let layout = self.variant_layouts.by_type.get(&tyname).cloned();
-            if let Some(layout) = layout {
-                // A GENERIC variant instantiates per-use (`Tree[Int]`): the
-                // helper key carries the args and the layout's field types are
-                // substituted, so `Leaf(T)` compares as `Leaf(Int)`.
-                let (key, layout) = instantiate_variant_layout(&tyname, &layout, ty);
-                if self.synth_eq_types.contains(&key) || self.variant_needs_eq_helper(&tyname) {
-                    if self.ensure_variant_eq_helper(&key, &layout) {
-                        let name = self.eq_helper_name(&key);
-                        return Some(self.emit_eq_helper_call(name, lv, rv));
-                    }
-                    return None;
-                }
-                return self.variant_eq_from_handles(lv, rv, &layout, depth);
-            }
-        }
-        // Option[T] — the scalar masked compare or the heap conditional compare.
-        if let Ty::Applied(TC::Option, oa) = ty {
-            if oa.len() == 1 {
-                if !is_heap_ty(&oa[0]) {
-                    return Some(self.option_scalar_eq_from_handles(lv, rv, &oa[0]));
-                }
-                return self.option_heap_eq_from_handles(lv, rv, &oa[0], depth);
-            }
-        }
-        // Result[T, E] — the proven masked core for the (scalar Ok, String Err) layout;
-        // the general both-Ok/both-Err conditional recursion for every other payload pair.
-        if let Ty::Applied(TC::Result, ra) = ty {
-            if ra.len() == 2 {
-                if !is_heap_ty(&ra[0]) && matches!(ra[1], Ty::String) {
-                    return self.result_scalar_eq_from_handles(lv, rv, &ra[0]);
-                }
-                return self.result_eq_general_from_handles(lv, rv, &ra[0], &ra[1], depth);
-            }
-        }
-        // Tuple / record — per-slot recursion, ANDed in declaration order.
-        if let Some((_names, ftys)) = self.aggregate_field_tys(ty) {
-            return self.aggregate_eq_from_handles(lv, rv, &ftys, depth);
-        }
-        None
-    }
-
-    /// Tuple/record `==` over two materialized block HANDLES: load each declaration-order
-    /// uniform slot (a heap field as its owned handle, a scalar as its value), recurse the
-    /// typed eq per field, AND-fold. An empty aggregate compares equal.
-    fn aggregate_eq_from_handles(
-        &mut self,
-        hl: ValueId,
-        hr: ValueId,
-        ftys: &[Ty],
-        depth: u32,
-    ) -> Option<ValueId> {
-        let mut acc: Option<ValueId> = None;
-        for (i, fty) in ftys.iter().enumerate() {
-            let off = crate::lower::layout::slot_offset(i) as i64;
-            let (lf, rf) = if is_heap_ty(fty) {
-                (self.load_payload_addr(hl, off), self.load_payload_addr(hr, off))
-            } else {
-                (
-                    self.load_at_offset(hl, off, crate::PrimKind::Load { width: 8 }),
-                    self.load_at_offset(hr, off, crate::PrimKind::Load { width: 8 }),
-                )
-            };
-            let e = self.typed_slot_eq(lf, rf, fty, depth + 1)?;
-            acc = Some(match acc {
-                None => e,
-                Some(prev) => {
-                    let dst = self.fresh_value();
-                    self.ops.push(Op::IntBinOp { dst, op: IntOp::And, a: prev, b: e });
-                    dst
-                }
-            });
-        }
-        Some(acc.unwrap_or_else(|| {
-            let dst = self.fresh_value();
-            self.ops.push(Op::ConstInt { dst, value: 1 });
-            dst
-        }))
-    }
-
-    /// Custom-variant `==` over two materialized block HANDLES (`[tag@slot0][field0@slot1]…`):
-    /// tag-eq AND a tag-dispatched field-compare chain — each field-carrying ctor's fields
-    /// recurse through the typed eq (ANDed), fieldless ctors compare true. All merge values are
-    /// scalar Bools — the IfThen/Else/EndIf merges carry no ownership. A field whose typed eq
-    /// cannot lower (e.g. a recursive `List[Self]` payload) propagates None → the caller walls.
-    pub(crate) fn variant_eq_from_handles(
-        &mut self,
-        hl: ValueId,
-        hr: ValueId,
-        layout: &crate::lower::VariantLayout,
-        depth: u32,
-    ) -> Option<ValueId> {
-        let toff = crate::lower::layout::slot_offset(0) as i64;
-        let tl = self.load_at_offset(hl, toff, crate::PrimKind::Load { width: 8 });
-        let tr = self.load_at_offset(hr, toff, crate::PrimKind::Load { width: 8 });
-        let t_eq = self.fresh_value();
-        self.ops.push(Op::IntBinOp { dst: t_eq, op: IntOp::Eq, a: tl, b: tr });
-        // Field chain under EQUAL tags: nested merges over the field-carrying ctors; the
-        // fieldless remainder compares true.
-        let fielded: Vec<(i64, Vec<Ty>)> = layout
-            .cases
-            .iter()
-            .filter(|c| !c.fields.is_empty())
-            .map(|c| (c.tag as i64, c.fields.iter().map(|(_, t)| t.clone()).collect()))
-            .collect();
-        let dst = self.fresh_value();
-        self.ops.push(Op::IfThen { cond: t_eq, dst: Some(dst) });
-        // then-branch: the chain value.
-        let mut ends: Vec<ValueId> = Vec::new();
-        for (tag, ftys) in &fielded {
-            let tagv = self.fresh_value();
-            self.ops.push(Op::ConstInt { dst: tagv, value: *tag });
-            let is_c = self.fresh_value();
-            self.ops.push(Op::IntBinOp { dst: is_c, op: IntOp::Eq, a: tl, b: tagv });
-            let d2 = self.fresh_value();
-            self.ops.push(Op::IfThen { cond: is_c, dst: Some(d2) });
-            // This ctor's fields at slots 1.. — recurse the typed eq per field, AND-fold.
-            let mut cmp: Option<ValueId> = None;
-            for (j, fty) in ftys.iter().enumerate() {
-                let foff = crate::lower::layout::slot_offset(1 + j) as i64;
-                let (l1, r1) = if is_heap_ty(fty) {
-                    (self.load_payload_addr(hl, foff), self.load_payload_addr(hr, foff))
-                } else {
-                    (
-                        self.load_at_offset(hl, foff, crate::PrimKind::Load { width: 8 }),
-                        self.load_at_offset(hr, foff, crate::PrimKind::Load { width: 8 }),
-                    )
-                };
-                let f_eq = self.typed_slot_eq(l1, r1, fty, depth + 1)?;
-                cmp = Some(match cmp {
-                    None => f_eq,
-                    Some(prev) => {
-                        let a2 = self.fresh_value();
-                        self.ops.push(Op::IntBinOp { dst: a2, op: IntOp::And, a: prev, b: f_eq });
-                        a2
-                    }
-                });
-            }
-            self.ops.push(Op::Else { val: Some(cmp?) });
-            ends.push(d2);
-        }
-        // innermost else: no field-ctor matched (a fieldless ctor) — equal.
-        let one = self.fresh_value();
-        self.ops.push(Op::ConstInt { dst: one, value: 1 });
-        let mut inner: ValueId = one;
-        // close the nested merges inside-out.
-        for d2 in ends.iter().rev() {
-            self.ops.push(Op::EndIf { val: Some(inner) });
-            inner = *d2;
-        }
-        // The chain's value is the OUTERMOST merge (`ends[0]`): each nested if's
-        // dst is only assigned along the path that reaches it, so yielding an
-        // inner dst reads an unassigned local whenever an OUTER arm was taken
-        // (first-ctor eq returned false with ≥2 fielded ctors).
-        let then_v = inner;
-        let zero = self.fresh_value();
-        self.ops.push(Op::Else { val: Some(then_v) });
-        self.ops.push(Op::ConstInt { dst: zero, value: 0 });
-        self.ops.push(Op::EndIf { val: Some(zero) });
-        Some(dst)
-    }
-
-    /// General `Result[T, E] ==` over two materialized block HANDLES (len@4 = 0 Ok / 1 Err,
-    /// payload@12) for payload pairs beyond the proven (scalar, String) masked core. Two SIBLING
-    /// gated compares — dst = (bothOk ? okEq : 0) | (bothErr ? errEq : 0) — so a payload slot is
-    /// only ever read INSIDE the branch that knows both sides carry that variant (a heap handle
-    /// is never dereferenced against the wrong layout), and a mixed Ok/Err pair compares false.
-    fn result_eq_general_from_handles(
-        &mut self,
-        hl: ValueId,
-        hr: ValueId,
-        ok_ty: &Ty,
-        err_ty: &Ty,
-        depth: u32,
-    ) -> Option<ValueId> {
-        // TWO Result layouts share this eq: a SCALAR-Ok Result is len-as-tag (@4: 0 = Ok,
-        // 1 = Err), but a HEAP-Ok Result is the cap-as-tag 1-slot block (len@4 is ALWAYS 1;
-        // the Ok/Err tag lives in slot 0's HIGH 32 bits @16 — materialize_result_str).
-        // Reading @4 on the heap-Ok layout classified EVERY value as Err, so
-        // `ok(xs) == ok(ys)` string-compared the two payload HANDLES (false for equal lists).
-        let tag_off: i64 = if is_heap_ty(ok_ty) { 16 } else { 4 };
-        let tag_l = self.load_at_offset(hl, tag_off, crate::PrimKind::Load { width: 4 });
-        let tag_r = self.load_at_offset(hr, tag_off, crate::PrimKind::Load { width: 4 });
-        let zero = self.fresh_value();
-        self.ops.push(Op::ConstInt { dst: zero, value: 0 });
-        let ok_l = self.fresh_value();
-        self.ops.push(Op::IntBinOp { dst: ok_l, op: IntOp::Eq, a: tag_l, b: zero });
-        let ok_r = self.fresh_value();
-        self.ops.push(Op::IntBinOp { dst: ok_r, op: IntOp::Eq, a: tag_r, b: zero });
-        let both_ok = self.fresh_value();
-        self.ops.push(Op::IntBinOp { dst: both_ok, op: IntOp::And, a: ok_l, b: ok_r });
-        let both_err = self.fresh_value();
-        self.ops.push(Op::IntBinOp { dst: both_err, op: IntOp::And, a: tag_l, b: tag_r });
-        // Gated Ok compare.
-        let ok_gate = self.fresh_value();
-        self.ops.push(Op::IfThen { cond: both_ok, dst: Some(ok_gate) });
-        let (pl, pr) = if is_heap_ty(ok_ty) {
-            (self.load_payload_addr(hl, 12), self.load_payload_addr(hr, 12))
-        } else {
-            (
-                self.load_at_offset(hl, 12, crate::PrimKind::Load { width: 8 }),
-                self.load_at_offset(hr, 12, crate::PrimKind::Load { width: 8 }),
-            )
-        };
-        let ok_eq = self.typed_slot_eq(pl, pr, ok_ty, depth + 1)?;
-        self.ops.push(Op::Else { val: Some(ok_eq) });
-        let f1 = self.fresh_value();
-        self.ops.push(Op::ConstInt { dst: f1, value: 0 });
-        self.ops.push(Op::EndIf { val: Some(f1) });
-        // Gated Err compare.
-        let err_gate = self.fresh_value();
-        self.ops.push(Op::IfThen { cond: both_err, dst: Some(err_gate) });
-        let (el, er) = if is_heap_ty(err_ty) {
-            (self.load_payload_addr(hl, 12), self.load_payload_addr(hr, 12))
-        } else {
-            (
-                self.load_at_offset(hl, 12, crate::PrimKind::Load { width: 8 }),
-                self.load_at_offset(hr, 12, crate::PrimKind::Load { width: 8 }),
-            )
-        };
-        let err_eq = self.typed_slot_eq(el, er, err_ty, depth + 1)?;
-        self.ops.push(Op::Else { val: Some(err_eq) });
-        let f2 = self.fresh_value();
-        self.ops.push(Op::ConstInt { dst: f2, value: 0 });
-        self.ops.push(Op::EndIf { val: Some(f2) });
-        let dst = self.fresh_value();
-        self.ops.push(Op::IntBinOp { dst, op: IntOp::Or, a: ok_gate, b: err_gate });
-        Some(dst)
-    }
-
-    /// The byte-address (`Prim::Handle`) of a materialized block — the operand handed to an eq core.
-    pub(crate) fn handle_of(&mut self, block: ValueId) -> ValueId {
-        let h = self.fresh_value();
-        self.ops.push(Op::Prim { kind: crate::PrimKind::Handle, dst: Some(h), args: vec![block] });
-        h
-    }
-
-    /// LoadHandle a nested payload slot AND normalize it to the eq engine's i64
-    /// byte-ADDRESS form (`Prim::Handle`): a raw `LoadHandle` result is an i32 Ptr
-    /// local, but every structural recursion (`load_at_offset`'s IntBinOp address
-    /// math) consumes handles as i64 — mixing the classes emitted
-    /// `(i64.add (local.get $v:i32))`, invalid wasm (the Option[(Int, String)] eq).
-    /// Module-eq callees (`string.eq`/`list.eq_*`) accept the i64 form through
-    /// `render_arg_wasm`'s Handle wrap, so normalizing is uniformly safe.
-    pub(crate) fn load_payload_addr(&mut self, h: ValueId, off: i64) -> ValueId {
-        let raw = self.load_at_offset(h, off, crate::PrimKind::LoadHandle);
-        self.handle_of(raw)
-    }
-
-    /// Materialize ONE operand of a heap `==` cond into a block whose handle the eq core reads.
-    /// - A tracked heap `Var` is BORROWED: its existing block is returned, NOT added to the cond
-    ///   frame (it is owned elsewhere and drops at its own scope — the eq only reads it).
-    /// - Any other heap operand (a heap-returning CALL, an Option/Result CTOR, a String literal /
-    ///   concat) is materialized into a FRESH OWNED block via `lower_owned_heap_field`, which
-    ///   pushes it to `live_heap_handles`; we then register its RECURSIVE drop set
-    ///   (`heap_elem_lists` / `value_handles` / …) so the cond-frame `drop_arm_locals` frees it
-    ///   AND its owned payload (no leak). The owned temp is freed exactly once (the frame `d`),
-    ///   never double-freed (the eq borrows it, never consumes). Returns None for a non-
-    ///   materializable shape → the caller walls.
-    fn materialize_eq_operand(&mut self, expr: &IrExpr, ty: &Ty) -> Option<ValueId> {
-        if let IrExprKind::Var { id } = &expr.kind {
-            // A tracked heap Var (a param or let-local) — borrow its block, no new ownership.
-            let v = self.value_for(*id).ok()?;
-            // The eq READS the block through the operand type's layout, so a DEFERRED
-            // Opaque bind (any heap type — a variant/tuple/record slot read of an empty
-            // block is a raw un-bounds-checked load, garbage compared silently) must
-            // decline on the strict path: the eq walls and v0 emits the correct bytes
-            // (the C-136 wrong-value family, consumer side).
-            if crate::lower::strict_values() && self.deferred_opaque_binds.contains(&v) {
-                return None;
-            }
-            // An Option/Result var must additionally be a genuinely MATERIALIZED block
-            // (ctor-built, or a param seeded by `seed_variant_param`). A deferred value
-            // that escaped the bind set — e.g. an optional-chain result outside the
-            // executable subset — has len 0, which the len-as-tag read MISREADS as
-            // `none`/`Ok` (`assert_eq(p?.name, some("Alice"))` compared false while
-            // native compared true — a silent wrong verdict, the #790 optional-chain
-            // row). Decline instead.
-            use almide_lang::types::constructor::TypeConstructorId as TC;
-            if matches!(ty, Ty::Applied(TC::Option | TC::Result, _))
-                && !self.materialized_options.contains(&v)
-                && !self.materialized_results.contains(&v)
-                && !self.materialized_results_str.contains(&v)
-            {
-                return None;
-            }
-            return Some(v);
-        }
-        // An Option/Result CTOR (`some("")`, `none`, `ok(v)`, `err(m)`) — `try_lower_option_ctor`
-        // (which handles BOTH Option and Result ctors) builds the DynListStr/OptSome block +
-        // registers its recursive drop set (`materialize_opt_str_some` → `heap_elem_lists` +
-        // `materialized_options`), but does NOT push to `live_heap_handles`. Push it so the cond-
-        // frame `drop_arm_locals` frees it (and its owned payload) exactly once after the
-        // (borrowing) eq.
-        if matches!(
-            &expr.kind,
-            IrExprKind::OptionSome { .. }
-                | IrExprKind::OptionNone
-                | IrExprKind::ResultOk { .. }
-                | IrExprKind::ResultErr { .. }
-        ) {
-            let obj = self.try_lower_option_ctor(expr, ty)?;
-            if !self.live_heap_handles.contains(&obj) {
-                self.live_heap_handles.push(obj);
-            }
-            return Some(obj);
-        }
-        // `rs[i] == ok(v)` — an ELEMENT of a materialized heap-element list (the
-        // fan.settle results literal): BORROW the element handle (`$elem_addr` +
-        // `LoadHandle` — the list owns it and the eq only reads), the Var-borrow
-        // discipline element-precise. Nothing joins the cond frame — no drop, no
-        // ownership event; an untracked/deferred container declines inside the
-        // borrow and falls through to the materializer below.
-        if let IrExprKind::IndexAccess { .. } = &expr.kind {
-            if let Some(b) = self.try_lower_heap_field_borrow(expr) {
-                return Some(b);
-            }
-        }
-        // Otherwise a heap-returning CALL / literal / concat — materialize a fresh OWNED block
-        // (`lower_owned_heap_field` pushes it to `live_heap_handles`). The call path leaves the
-        // block FLAT, so register the recursive drop set from the operand TYPE — else an
-        // `Option[String]`/`List[String]` temp leaks its inner Strings. Idempotent.
-        let obj = self.lower_owned_heap_field(expr)?;
-        if self.live_heap_handles.contains(&obj) {
-            self.register_owned_heap_eq_drop(obj, ty);
-        }
-        Some(obj)
-    }
-
-    /// Register the recursive drop set for a freshly materialized heap eq-operand block, mirroring
-    /// the call-binding tracking in `lower_bind` so the cond-frame teardown frees nested ownership.
-    pub(crate) fn register_owned_heap_eq_drop(&mut self, obj: ValueId, ty: &Ty) {
-        if crate::lower::is_list_list_str_ty(ty) {
-            self.list_list_str_lists.insert(obj);
-        } else if crate::lower::is_list_str_str_ty(ty) {
-            self.str_str_elem_lists.insert(obj);
-        } else if crate::lower::is_list_int_str_ty(ty) {
-            self.variant_drop_handles.insert(obj, "list_int_str".to_string());
-        } else if crate::lower::is_lenlist_list_ty(ty) {
-            self.variant_drop_handles.insert(obj, "list_lenlist".to_string());
-        } else if is_heap_elem_list_ty(ty) {
-            // List[heap] / Option[heap] / Result[_, heap] — the DynListStr recursive free.
-            self.heap_elem_lists.insert(obj);
-        }
-        if crate::lower::is_value_ty(ty) {
-            self.value_handles.insert(obj);
-        }
-    }
-
-    fn lower_heap_result_if_inner(
-        &mut self,
-        cond: &IrExpr,
-        then: &IrExpr,
-        else_: &IrExpr,
-        result_ty: &Ty,
-    ) -> Option<ValueId> {
-        let cond_v = self.lower_heap_result_cond(cond)?;
-        let dst = self.fresh_value();
-        self.ops.push(Op::IfThen { cond: cond_v, dst: Some(dst) });
-        // RELEASE PARITY across the arms. An arm may MOVE an OUTER scope-level
-        // handle into its result — the effect tail's `err(msg)` moves the error
-        // accumulator into the Err block — which removes it from
-        // `live_heap_handles` GLOBALLY, though the move runs only on that arm's
-        // PATH. The sibling arm must then release it ITSELF: without the
-        // compensating Drop the non-moving path LEAKS the handle (one error
-        // accumulator per call on the happy path). The flat certificate hid this
-        // by counting the moving arm's `m` unconditionally; the branch-grouped
-        // cert (`{m|}` — arms disagree) REJECTS it, which is how it was found.
-        // With the Drop the arms agree (`{m|d}`) and the leak is gone. Nested
-        // heap-result ifs recurse through here, so parity holds level by level.
-        let outer: Vec<ValueId> = self.live_heap_handles.clone();
-        let then_obj = self.lower_heap_result_arm(then, result_ty)?;
-        let consumed_by_then: Vec<ValueId> =
-            outer.iter().copied().filter(|h| !self.live_heap_handles.contains(h)).collect();
-        let else_marker_at = self.ops.len();
-        self.ops.push(Op::Else { val: Some(then_obj) });
-        let live_after_then: Vec<ValueId> = self.live_heap_handles.clone();
-        let else_obj = self.lower_heap_result_arm(else_, result_ty)?;
-        let consumed_by_else: Vec<ValueId> = live_after_then
-            .iter()
-            .copied()
-            .filter(|h| !self.live_heap_handles.contains(h))
-            .collect();
-        for h in &consumed_by_then {
-            if !consumed_by_else.contains(h) {
-                let op = self.drop_op_for(*h);
-                self.ops.push(op); // the ELSE arm releases what THEN moved out
-            }
-        }
-        for h in &consumed_by_else {
-            if !consumed_by_then.contains(h) {
-                let op = self.drop_op_for(*h);
-                self.ops.insert(else_marker_at, op); // the THEN arm releases what ELSE moved out
-            }
-        }
-        self.ops.push(Op::EndIf { val: Some(else_obj) });
-        Some(dst)
-    }
 }
